refactor(store): migrate SQL to sqlc and split store into domain files
Replace manual SQL with sqlc v1.31.1 typed queries. All operational SQL
now lives in internal/db/queries/ and is generated into internal/db/generated/.
Dynamic SQL concatenation eliminated; allowlist-guarded DDL migrations remain.

Split the 3600-line store.go monolith into domain files: session, observation,
tag, sync, project, export, passive, helpers, mappers, types. Public API unchanged.

CI added: sqlc generation drift check, SQL concatenation guard, race tests.
