+++ internal/store/helpers.go
@@ -1,0 +1,193 @@
+package store
+
+import (
+	"crypto/rand"
+	"crypto/sha256"
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"regexp"
+	"strconv"
+	"strings"
+	"time"
+)
+
+// nonAlnumRe matches one or more non-alphanumeric characters.
+var nonAlnumRe = regexp.MustCompile(`[^a-z0-9]+`)
+
+// privateTagRegex matches <private>...</private> tags and their contents.
+var privateTagRegex = regexp.MustCompile(`(?is)<private>.*?</private>`)
+
+func stripPrivateTags(s string) string {
+	result := privateTagRegex.ReplaceAllString(s, "[REDACTED]")
+	return strings.TrimSpace(result)
+}
+
+// sanitizeFTS wraps each word in quotes so FTS5 doesn't choke on special chars.
+func sanitizeFTS(query string) string {
+	words := strings.Fields(query)
+	for i, w := range words {
+		w = strings.Trim(w, `"`)
+		words[i] = `"` + w + `"`
+	}
+	return strings.Join(words, " OR ")
+}
+
+func hashNormalized(content string) string {
+	normalized := strings.ToLower(strings.Join(strings.Fields(content), " "))
+	h := sha256.Sum256([]byte(normalized))
+	return hex.EncodeToString(h[:])
+}
+
+func normalizeScope(scope string) string {
+	v := strings.TrimSpace(strings.ToLower(scope))
+	if v == "personal" {
+		return "personal"
+	}
+	return "project"
+}
+
+func normalizeOptionalScope(scope string) string {
+	if scope == "" {
+		return ""
+	}
+	return normalizeScope(scope)
+}
+
+func normalizeTopicKey(topic string) string {
+	v := strings.TrimSpace(strings.ToLower(topic))
+	if v == "" {
+		return ""
+	}
+	v = strings.Join(strings.Fields(v), "-")
+	if len(v) > 120 {
+		v = v[:120]
+	}
+	return v
+}
+
+func normalizeTopicSegment(s string) string {
+	v := strings.ToLower(strings.TrimSpace(s))
+	if v == "" {
+		return ""
+	}
+	v = nonAlnumRe.ReplaceAllString(v, " ")
+	v = strings.Join(strings.Fields(v), "-")
+	if len(v) > 100 {
+		v = v[:100]
+	}
+	return v
+}
+
+func truncate(s string, max int) string {
+	runes := []rune(s)
+	if len(runes) <= max {
+		return s
+	}
+	return string(runes[:max]) + "..."
+}
+
+func derefString(v *string) string {
+	if v == nil {
+		return ""
+	}
+	return *v
+}
+
+func dedupeWindowExpression(window time.Duration) string {
+	if window <= 0 {
+		window = 15 * time.Minute
+	}
+	minutes := int(window.Minutes())
+	if minutes < 1 {
+		minutes = 1
+	}
+	return "-" + strconv.Itoa(minutes) + " minutes"
+}
+
+func maxInt(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}
+
+func normalizeSyncTargetKey(targetKey string) string {
+	if strings.TrimSpace(targetKey) == "" {
+		return DefaultSyncTargetKey
+	}
+	return strings.TrimSpace(strings.ToLower(targetKey))
+}
+
+func newSyncID(prefix string) string {
+	b := make([]byte, 8)
+	if _, err := rand.Read(b); err != nil {
+		return fmt.Sprintf("%s-%d", prefix, time.Now().UTC().UnixNano())
+	}
+	return prefix + "-" + hex.EncodeToString(b)
+}
+
+func normalizeExistingSyncID(existing, prefix string) string {
+	if strings.TrimSpace(existing) != "" {
+		return existing
+	}
+	return newSyncID(prefix)
+}
+
+func extractProjectFromPayload(payload any) string {
+	switch p := payload.(type) {
+	case syncSessionPayload:
+		return p.Project
+	case syncObservationPayload:
+		if p.Project != nil {
+			return *p.Project
+		}
+		return ""
+	case syncPromptPayload:
+		if p.Project != nil {
+			return *p.Project
+		}
+		return ""
+	default:
+		data, err := json.Marshal(payload)
+		if err != nil {
+			return ""
+		}
+		var generic struct {
+			Project *string `json:"project"`
+		}
+		if err := json.Unmarshal(data, &generic); err != nil || generic.Project == nil {
+			return ""
+		}
+		return *generic.Project
+	}
+}
+
+func decodeSyncPayload(payload []byte, dest any) error {
+	trimmed := strings.TrimSpace(string(payload))
+	if trimmed == "" {
+		return fmt.Errorf("empty payload")
+	}
+	if trimmed[0] != '"' {
+		return json.Unmarshal([]byte(trimmed), dest)
+	}
+	var encoded string
+	if err := json.Unmarshal([]byte(trimmed), &encoded); err != nil {
+		return err
+	}
+	return json.Unmarshal([]byte(encoded), dest)
+}
+
+func hasAny(text string, words ...string) bool {
+	for _, w := range words {
+		if strings.Contains(text, w) {
+			return true
+		}
+	}
+	return false
+}
+
+// Now returns the current time formatted for SQLite.
+func Now() string {
+	return time.Now().UTC().Format("2006-01-02 15:04:05")
+}
