+++ internal/store/store.go
@@ -6,312 +6,26 @@
 package store
 
 import (
-	"crypto/rand"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
 	"regexp"
-	"sort"
-	"strconv"
-	"strings"
 	"time"
 
+	dbgen "github.com/jmeiracorbal/mnemo/internal/db/generated"
 	_ "modernc.org/sqlite"
 )
 
 var openDB = sql.Open
 
-// ─── Types ───────────────────────────────────────────────────────────────────
+var sqlIdentifierPattern = regexp.MustCompile(`^[A-Za-z_][A-Za-z0-9_]*$`)
 
-type Project struct {
-	ID        string `json:"id"`
-	Name      string `json:"name"`
-	CreatedAt string `json:"created_at"`
-}
-
-type Session struct {
-	ID        string   `json:"id"`
-	Project   string   `json:"project"`
-	Directory string   `json:"directory"`
-	StartedAt string   `json:"started_at"`
-	EndedAt   *string  `json:"ended_at,omitempty"`
-	Summary   *string  `json:"summary,omitempty"`
-	Tags      []string `json:"tags,omitempty"`
-}
-
-type Observation struct {
-	ID             int64    `json:"id"`
-	SyncID         string   `json:"sync_id"`
-	SessionID      string   `json:"session_id"`
-	Type           string   `json:"type"`
-	Title          string   `json:"title"`
-	Content        string   `json:"content"`
-	ToolName       *string  `json:"tool_name,omitempty"`
-	Project        *string  `json:"project,omitempty"`
-	Scope          string   `json:"scope"`
-	TopicKey       *string  `json:"topic_key,omitempty"`
-	Tags           []string `json:"tags,omitempty"`
-	RevisionCount  int      `json:"revision_count"`
-	DuplicateCount int      `json:"duplicate_count"`
-	LastSeenAt     *string  `json:"last_seen_at,omitempty"`
-	CreatedAt      string   `json:"created_at"`
-	UpdatedAt      string   `json:"updated_at"`
-	DeletedAt      *string  `json:"deleted_at,omitempty"`
-}
-
-type TagInfo struct {
-	Tag        string `json:"tag"`
-	Count      int    `json:"count"`
-	LastUsedAt string `json:"last_used_at"`
-}
-
-// TagStatsOptions controls filtering and ordering in TagStats.
-type TagStatsOptions struct {
-	// MinCount includes only tags used at least this many times. Zero means no lower bound.
-	MinCount int
-	// MaxCount includes only tags used at most this many times. Zero means no upper bound.
-	MaxCount int
-	// UnusedSince includes only tags whose last use predates this time. Zero means no filter.
-	UnusedSince time.Time
-	// Limit caps the number of results. Zero means no limit.
-	Limit int
-	// SortBy controls result ordering: "freq" (default, highest first),
-	// "stale" (oldest last_used_at first), or "alpha" (alphabetical).
-	SortBy string
-}
-
-// TagWeight holds observability signals for a single tag.
-// Prepared for use in future retrieval weighting — not yet wired into any ranking path.
-type TagWeight struct {
-	Tag          string
-	Frequency    int
-	RecencyScore float64 // 0.0 (oldest) to 1.0 (most recent), relative to the set
-	IsDominant   bool    // true if this tag is within the top N by frequency
-}
-
-// RelatedTag represents a tag that co-occurs with a queried tag.
-type RelatedTag struct {
-	Tag               string  `json:"tag"`
-	CooccurrenceCount int     `json:"cooccurrence_count"`
-	Score             float64 `json:"score"`
-	LastSeenAt        string  `json:"last_seen_at"`
-}
-
-// RelatedTagsOptions controls RelatedTags behaviour.
-type RelatedTagsOptions struct {
-	// Limit caps results. Zero means no limit.
-	Limit int
-	// Since restricts to co-occurrences seen after this time. Zero means no filter.
-	Since time.Time
-	// MinCooccurrence filters out tags that co-occur fewer than this many times. Zero means no filter.
-	MinCooccurrence int
-	// IncludeObservations counts co-occurrences from observation_tags (default: true when both are false).
-	IncludeObservations bool
-	// IncludeSessions counts co-occurrences from session_tags (default: true when both are false).
-	IncludeSessions bool
-}
-
-type SearchResult struct {
-	Observation
-	Rank float64 `json:"rank"`
-}
-
-type SessionSummary struct {
-	ID               string  `json:"id"`
-	Project          string  `json:"project"`
-	StartedAt        string  `json:"started_at"`
-	EndedAt          *string `json:"ended_at,omitempty"`
-	Summary          *string `json:"summary,omitempty"`
-	ObservationCount int     `json:"observation_count"`
-}
-
-type Stats struct {
-	TotalSessions     int      `json:"total_sessions"`
-	TotalObservations int      `json:"total_observations"`
-	TotalPrompts      int      `json:"total_prompts"`
-	Projects          []string `json:"projects"`
-}
-
-type TimelineEntry struct {
-	ID             int64   `json:"id"`
-	SessionID      string  `json:"session_id"`
-	Type           string  `json:"type"`
-	Title          string  `json:"title"`
-	Content        string  `json:"content"`
-	ToolName       *string `json:"tool_name,omitempty"`
-	Project        *string `json:"project,omitempty"`
-	Scope          string  `json:"scope"`
-	TopicKey       *string `json:"topic_key,omitempty"`
-	RevisionCount  int     `json:"revision_count"`
-	DuplicateCount int     `json:"duplicate_count"`
-	LastSeenAt     *string `json:"last_seen_at,omitempty"`
-	CreatedAt      string  `json:"created_at"`
-	UpdatedAt      string  `json:"updated_at"`
-	DeletedAt      *string `json:"deleted_at,omitempty"`
-	IsFocus        bool    `json:"is_focus"` // true for the anchor observation
-}
-
-type TimelineResult struct {
-	Focus        Observation     `json:"focus"`        // The anchor observation
-	Before       []TimelineEntry `json:"before"`       // Observations before the focus (chronological)
-	After        []TimelineEntry `json:"after"`        // Observations after the focus (chronological)
-	SessionInfo  *Session        `json:"session_info"` // Session that contains the focus observation
-	TotalInRange int             `json:"total_in_range"`
-}
-
-type SearchOptions struct {
-	Type      string   `json:"type,omitempty"`
-	Project   string   `json:"project,omitempty"`
-	Scope     string   `json:"scope,omitempty"`
-	TopicKey  string   `json:"topic_key,omitempty"`
-	Tags      []string `json:"tags,omitempty"`
-	PreferTags []string `json:"prefer_tags,omitempty"`
-	Limit     int      `json:"limit,omitempty"`
-}
-
-// ContextOptions controls retrieval behaviour in RecentObservations / FormatContext.
-type ContextOptions struct {
-	// Tags is a hard filter: only observations with ALL listed tags are included.
-	Tags []string
-	// PreferTags is a soft signal: observations matching more of these tags rank higher.
-	PreferTags []string
-	// TopicKey boosts observations that share this topic_key without excluding others.
-	TopicKey string
-}
-
-type AddObservationParams struct {
-	SessionID string   `json:"session_id"`
-	Type      string   `json:"type"`
-	Title     string   `json:"title"`
-	Content   string   `json:"content"`
-	ToolName  string   `json:"tool_name,omitempty"`
-	Project   string   `json:"project,omitempty"`
-	Scope     string   `json:"scope,omitempty"`
-	TopicKey  string   `json:"topic_key,omitempty"`
-	Tags      []string `json:"tags,omitempty"`
-}
-
-type UpdateObservationParams struct {
-	Type     *string   `json:"type,omitempty"`
-	Title    *string   `json:"title,omitempty"`
-	Content  *string   `json:"content,omitempty"`
-	Project  *string   `json:"project,omitempty"`
-	Scope    *string   `json:"scope,omitempty"`
-	TopicKey *string   `json:"topic_key,omitempty"`
-	Tags     *[]string `json:"tags,omitempty"`
-}
-
-type Prompt struct {
-	ID        int64  `json:"id"`
-	SyncID    string `json:"sync_id"`
-	SessionID string `json:"session_id"`
-	Content   string `json:"content"`
-	Project   string `json:"project,omitempty"`
-	CreatedAt string `json:"created_at"`
-}
-
-type AddPromptParams struct {
-	SessionID string `json:"session_id"`
-	Content   string `json:"content"`
-	Project   string `json:"project,omitempty"`
-}
-
-const (
-	DefaultSyncTargetKey = "cloud"
-
-	SyncLifecycleIdle     = "idle"
-	SyncLifecyclePending  = "pending"
-	SyncLifecycleRunning  = "running"
-	SyncLifecycleHealthy  = "healthy"
-	SyncLifecycleDegraded = "degraded"
-
-	SyncEntitySession     = "session"
-	SyncEntityObservation = "observation"
-	SyncEntityPrompt      = "prompt"
-
-	SyncOpUpsert = "upsert"
-	SyncOpDelete = "delete"
-
-	SyncSourceLocal  = "local"
-	SyncSourceRemote = "remote"
-)
-
-type SyncState struct {
-	TargetKey           string  `json:"target_key"`
-	Lifecycle           string  `json:"lifecycle"`
-	LastEnqueuedSeq     int64   `json:"last_enqueued_seq"`
-	LastAckedSeq        int64   `json:"last_acked_seq"`
-	LastPulledSeq       int64   `json:"last_pulled_seq"`
-	ConsecutiveFailures int     `json:"consecutive_failures"`
-	BackoffUntil        *string `json:"backoff_until,omitempty"`
-	LeaseOwner          *string `json:"lease_owner,omitempty"`
-	LeaseUntil          *string `json:"lease_until,omitempty"`
-	LastError           *string `json:"last_error,omitempty"`
-	UpdatedAt           string  `json:"updated_at"`
-}
-
-type SyncMutation struct {
-	Seq        int64   `json:"seq"`
-	TargetKey  string  `json:"target_key"`
-	Entity     string  `json:"entity"`
-	EntityKey  string  `json:"entity_key"`
-	Op         string  `json:"op"`
-	Payload    string  `json:"payload"`
-	Source     string  `json:"source"`
-	Project    string  `json:"project"`
-	OccurredAt string  `json:"occurred_at"`
-	AckedAt    *string `json:"acked_at,omitempty"`
-}
-
-// EnrolledProject represents a project enrolled for cloud sync.
-type EnrolledProject struct {
-	Project    string `json:"project"`
-	EnrolledAt string `json:"enrolled_at"`
-}
-
-type syncSessionPayload struct {
-	ID        string    `json:"id"`
-	Project   string    `json:"project"`
-	Directory string    `json:"directory"`
-	EndedAt   *string   `json:"ended_at,omitempty"`
-	Summary   *string   `json:"summary,omitempty"`
-	Tags      *[]string `json:"tags,omitempty"`
-}
-
-type syncObservationPayload struct {
-	SyncID     string    `json:"sync_id"`
-	SessionID  string    `json:"session_id"`
-	Type       string    `json:"type"`
-	Title      string    `json:"title"`
-	Content    string    `json:"content"`
-	ToolName   *string   `json:"tool_name,omitempty"`
-	Project    *string   `json:"project,omitempty"`
-	Scope      string    `json:"scope"`
-	TopicKey   *string   `json:"topic_key,omitempty"`
-	Tags       *[]string `json:"tags,omitempty"`
-	Deleted    bool      `json:"deleted,omitempty"`
-	DeletedAt  *string   `json:"deleted_at,omitempty"`
-	HardDelete bool      `json:"hard_delete,omitempty"`
-}
-
-type syncPromptPayload struct {
-	SyncID    string  `json:"sync_id"`
-	SessionID string  `json:"session_id"`
-	Content   string  `json:"content"`
-	Project   *string `json:"project,omitempty"`
-}
-
-// ExportData is the full serializable dump of the mnemo database.
-type ExportData struct {
-	Version      string        `json:"version"`
-	ExportedAt   string        `json:"exported_at"`
-	Sessions     []Session     `json:"sessions"`
-	Observations []Observation `json:"observations"`
-	Prompts      []Prompt      `json:"prompts"`
+var allowedColumnDefinitions = map[string]bool{
+	"TEXT":                            true,
+	"TEXT NOT NULL DEFAULT 'project'": true,
+	"INTEGER NOT NULL DEFAULT 1":      true,
+	"TEXT NOT NULL DEFAULT ''":        true,
 }
 
 // ─── Config ──────────────────────────────────────────────────────────────────
@@ -339,8 +53,6 @@ func DefaultConfig() (Config, error) {
 }
 
 // FallbackConfig returns a Config with the given DataDir and default values.
-// Use this when DefaultConfig fails and you have resolved the home directory
-// through alternative means.
 func FallbackConfig(dataDir string) Config {
 	return Config{
 		DataDir:              dataDir,
@@ -360,6 +72,7 @@ func (s *Store) MaxObservationLength() int {
 
 type Store struct {
 	db    *sql.DB
+	q     *dbgen.Queries
 	cfg   Config
 	hooks storeHooks
 }
@@ -383,21 +96,10 @@ type sqlRowScanner struct {
 	rows *sql.Rows
 }
 
-func (r sqlRowScanner) Next() bool {
-	return r.rows.Next()
-}
-
-func (r sqlRowScanner) Scan(dest ...any) error {
-	return r.rows.Scan(dest...)
-}
-
-func (r sqlRowScanner) Err() error {
-	return r.rows.Err()
-}
-
-func (r sqlRowScanner) Close() error {
-	return r.rows.Close()
-}
+func (r sqlRowScanner) Next() bool        { return r.rows.Next() }
+func (r sqlRowScanner) Scan(dest ...any) error { return r.rows.Scan(dest...) }
+func (r sqlRowScanner) Err() error        { return r.rows.Err() }
+func (r sqlRowScanner) Close() error      { return r.rows.Close() }
 
 type storeHooks struct {
 	exec    func(db execer, query string, args ...any) (sql.Result, error)
@@ -484,7 +186,6 @@ func New(cfg Config) (*Store, error) {
 		return nil, fmt.Errorf("mnemo: open database: %w", err)
 	}
 
-	// SQLite performance pragmas
 	pragmas := []string{
 		"PRAGMA journal_mode = WAL",
 		"PRAGMA busy_timeout = 5000",
@@ -497,11 +198,13 @@ func New(cfg Config) (*Store, error) {
 		}
 	}
 
-	s := &Store{db: db, cfg: cfg, hooks: defaultStoreHooks()}
+	s := &Store{db: db, q: dbgen.New(db), cfg: cfg, hooks: defaultStoreHooks()}
 	if err := s.migrate(); err != nil {
+		_ = db.Close()
 		return nil, fmt.Errorf("mnemo: migration: %w", err)
 	}
 	if err := s.repairEnrolledProjectSyncMutations(); err != nil {
+		_ = db.Close()
 		return nil, fmt.Errorf("mnemo: repair enrolled sync journal: %w", err)
 	}
 
@@ -659,7 +362,6 @@ func (s *Store) migrate() error {
 		return err
 	}
 
-	// Project-scoped sync: add project column to sync_mutations and enrollment table.
 	if err := s.addColumnIfNotExists("sync_mutations", "project", "TEXT NOT NULL DEFAULT ''"); err != nil {
 		return err
 	}
@@ -672,7 +374,6 @@ func (s *Store) migrate() error {
 	`); err != nil {
 		return err
 	}
-	// Backfill: extract project from JSON payload for existing rows with empty project.
 	if _, err := s.execHook(s.db, `
 		UPDATE sync_mutations
 		SET project = COALESCE(json_extract(payload, '$.project'), '')
@@ -745,7 +446,6 @@ func (s *Store) migrate() error {
 	return nil
 }
 
-// ensureFTSTriggers creates FTS5 sync triggers using IF NOT EXISTS — idempotent.
 func (s *Store) ensureFTSTriggers() error {
 	_, err := s.execHook(s.db, `
 		CREATE TRIGGER IF NOT EXISTS obs_fts_insert AFTER INSERT ON observations BEGIN
@@ -785,2691 +485,13 @@ func (s *Store) ensureFTSTriggers() error {
 	return err
 }
 
-// ─── Projects ────────────────────────────────────────────────────────────────
-
-// EnsureProject inserts a project row if it does not already exist. Idempotent.
-func (s *Store) EnsureProject(id, name string) error {
-	_, err := s.execHook(s.db,
-		`INSERT OR IGNORE INTO projects (id, name) VALUES (?, ?)`,
-		id, name,
-	)
-	return err
-}
-
-// GetProjectByID returns the project with the given ID, or nil if not found.
-func (s *Store) GetProjectByID(id string) (*Project, error) {
-	row := s.db.QueryRow(
-		`SELECT id, name, created_at FROM projects WHERE id = ?`, id,
-	)
-	var p Project
-	if err := row.Scan(&p.ID, &p.Name, &p.CreatedAt); err == sql.ErrNoRows {
-		return nil, nil
-	} else if err != nil {
-		return nil, err
-	}
-	return &p, nil
-}
-
-// ListProjects returns all projects ordered by creation time.
-func (s *Store) ListProjects() ([]Project, error) {
-	rows, err := s.queryItHook(s.db,
-		`SELECT id, name, created_at FROM projects ORDER BY created_at ASC`,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var projects []Project
-	for rows.Next() {
-		var p Project
-		if err := rows.Scan(&p.ID, &p.Name, &p.CreatedAt); err != nil {
-			return nil, err
-		}
-		projects = append(projects, p)
-	}
-	return projects, rows.Err()
-}
-
-// ─── Sessions ────────────────────────────────────────────────────────────────
-
-func (s *Store) CreateSession(id, project, directory string) error {
-	return s.withTx(func(tx *sql.Tx) error {
-		if err := s.createSessionTx(tx, id, project, directory); err != nil {
-			return err
-		}
-		return s.enqueueSyncMutationTx(tx, SyncEntitySession, id, SyncOpUpsert, syncSessionPayload{
-			ID:        id,
-			Project:   project,
-			Directory: directory,
-		})
-	})
-}
-
-func (s *Store) EndSession(id string, summary string) error {
-	return s.withTx(func(tx *sql.Tx) error {
-		res, err := s.execHook(tx,
-			`UPDATE sessions SET ended_at = datetime('now'), summary = ? WHERE id = ?`,
-			nullableString(summary), id,
-		)
-		if err != nil {
-			return err
-		}
-		rows, err := res.RowsAffected()
-		if err != nil {
-			return err
-		}
-		if rows == 0 {
-			return nil
-		}
-
-		var endedAt string
-		var project, directory string
-		var storedSummary *string
-		if err := tx.QueryRow(
-			`SELECT project, directory, ended_at, summary FROM sessions WHERE id = ?`,
-			id,
-		).Scan(&project, &directory, &endedAt, &storedSummary); err != nil {
-			return err
-		}
-
-		return s.enqueueSyncMutationTx(tx, SyncEntitySession, id, SyncOpUpsert, syncSessionPayload{
-			ID:        id,
-			Project:   project,
-			Directory: directory,
-			EndedAt:   &endedAt,
-			Summary:   storedSummary,
-		})
-	})
-}
-
-func (s *Store) GetSession(id string) (*Session, error) {
-	row := s.db.QueryRow(
-		`SELECT id, project, directory, started_at, ended_at, summary FROM sessions WHERE id = ?`, id,
-	)
-	var sess Session
-	if err := row.Scan(&sess.ID, &sess.Project, &sess.Directory, &sess.StartedAt, &sess.EndedAt, &sess.Summary); err != nil {
-		return nil, err
-	}
-	s.loadTagsForSession(&sess)
-	return &sess, nil
-}
-
-// ObsCount returns the number of non-deleted observations for a given session.
-func (s *Store) ObsCount(sessionID string) (int, error) {
-	var count int
-	err := s.db.QueryRow(
-		`SELECT COUNT(*) FROM observations WHERE session_id = ? AND deleted_at IS NULL`, sessionID,
-	).Scan(&count)
-	return count, err
-}
-
-// ObsCountForSession counts all non-deleted observations saved for the session's
-// project on or after the session started. This catches observations saved via MCP
-// tools with a different session_id than the system session_id.
-func (s *Store) ObsCountForSession(sessionID string) (int, error) {
-	var count int
-	err := s.db.QueryRow(`
-		SELECT COUNT(*) FROM observations o
-		JOIN sessions ss ON ss.id = ?
-		WHERE o.project = ss.project
-		  AND o.created_at >= ss.started_at
-		  AND o.deleted_at IS NULL`,
-		sessionID,
-	).Scan(&count)
-	return count, err
-}
-
-func (s *Store) RecentSessions(project string, limit int) ([]SessionSummary, error) {
-	if limit <= 0 {
-		limit = 5
-	}
-
-	query := `
-		SELECT s.id, s.project, s.started_at, s.ended_at, s.summary,
-		       COUNT(o.id) as observation_count
-		FROM sessions s
-		LEFT JOIN observations o ON o.session_id = s.id AND o.deleted_at IS NULL
-		WHERE 1=1
-	`
-	args := []any{}
-
-	if project != "" {
-		query += " AND s.project = ?"
-		args = append(args, project)
-	}
-
-	query += " GROUP BY s.id ORDER BY MAX(COALESCE(o.created_at, s.started_at)) DESC LIMIT ?"
-	args = append(args, limit)
-
-	rows, err := s.queryItHook(s.db, query, args...)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var results []SessionSummary
-	for rows.Next() {
-		var ss SessionSummary
-		if err := rows.Scan(&ss.ID, &ss.Project, &ss.StartedAt, &ss.EndedAt, &ss.Summary, &ss.ObservationCount); err != nil {
-			return nil, err
-		}
-		results = append(results, ss)
-	}
-	return results, rows.Err()
-}
-
-// AllSessions returns recent sessions ordered by most recent first (for TUI browsing).
-func (s *Store) AllSessions(project string, limit int) ([]SessionSummary, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-
-	query := `
-		SELECT s.id, s.project, s.started_at, s.ended_at, s.summary,
-		       COUNT(o.id) as observation_count
-		FROM sessions s
-		LEFT JOIN observations o ON o.session_id = s.id AND o.deleted_at IS NULL
-		WHERE 1=1
-	`
-	args := []any{}
-
-	if project != "" {
-		query += " AND s.project = ?"
-		args = append(args, project)
-	}
-
-	query += " GROUP BY s.id ORDER BY MAX(COALESCE(o.created_at, s.started_at)) DESC LIMIT ?"
-	args = append(args, limit)
-
-	rows, err := s.queryItHook(s.db, query, args...)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var results []SessionSummary
-	for rows.Next() {
-		var ss SessionSummary
-		if err := rows.Scan(&ss.ID, &ss.Project, &ss.StartedAt, &ss.EndedAt, &ss.Summary, &ss.ObservationCount); err != nil {
-			return nil, err
-		}
-		results = append(results, ss)
-	}
-	return results, rows.Err()
-}
-
-// AllObservations returns recent observations ordered by most recent first (for TUI browsing).
-// Optional tags parameter filters to observations that have ALL specified tags.
-func (s *Store) AllObservations(project, scope string, limit int, tags ...string) ([]Observation, error) {
-	if limit <= 0 {
-		limit = s.cfg.MaxContextResults
-	}
-
-	query := `
-		SELECT o.id, ifnull(o.sync_id, '') as sync_id, o.session_id, o.type, o.title, o.content, o.tool_name, o.project,
-		       o.scope, o.topic_key, o.revision_count, o.duplicate_count, o.last_seen_at, o.created_at, o.updated_at, o.deleted_at
-		FROM observations o
-		WHERE o.deleted_at IS NULL
-	`
-	args := []any{}
-
-	if project != "" {
-		query += " AND o.project = ?"
-		args = append(args, project)
-	}
-	if scope != "" {
-		query += " AND o.scope = ?"
-		args = append(args, normalizeScope(scope))
-	}
-	for _, tag := range tags {
-		if norm := normalizeTag(tag); norm != "" {
-			query += " AND o.id IN (SELECT observation_id FROM observation_tags WHERE tag = ?)"
-			args = append(args, norm)
-		}
-	}
-
-	query += " ORDER BY o.created_at DESC LIMIT ?"
-	args = append(args, limit)
-
-	return s.queryObservations(query, args...)
-}
-
-// SessionObservations returns all observations for a specific session.
-func (s *Store) SessionObservations(sessionID string, limit int) ([]Observation, error) {
-	if limit <= 0 {
-		limit = 200
-	}
-
-	query := `
-		SELECT id, ifnull(sync_id, '') as sync_id, session_id, type, title, content, tool_name, project,
-		       scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		FROM observations
-		WHERE session_id = ? AND deleted_at IS NULL
-		ORDER BY created_at ASC
-		LIMIT ?
-	`
-	return s.queryObservations(query, sessionID, limit)
-}
-
-// ─── Observations ────────────────────────────────────────────────────────────
-
-func (s *Store) AddObservation(p AddObservationParams) (int64, error) {
-	// Strip <private>...</private> tags before persisting ANYTHING
-	title := stripPrivateTags(p.Title)
-	content := stripPrivateTags(p.Content)
-
-	if len(content) > s.cfg.MaxObservationLength {
-		content = content[:s.cfg.MaxObservationLength] + "... [truncated]"
-	}
-	scope := normalizeScope(p.Scope)
-	normHash := hashNormalized(content)
-	topicKey := normalizeTopicKey(p.TopicKey)
-
-	var observationID int64
-	err := s.withTx(func(tx *sql.Tx) error {
-		var obs *Observation
-		if topicKey != "" {
-			var existingID int64
-			err := tx.QueryRow(
-				`SELECT id FROM observations
-				 WHERE topic_key = ?
-				   AND ifnull(project, '') = ifnull(?, '')
-				   AND scope = ?
-				   AND deleted_at IS NULL
-				 ORDER BY datetime(updated_at) DESC, datetime(created_at) DESC
-				 LIMIT 1`,
-				topicKey, nullableString(p.Project), scope,
-			).Scan(&existingID)
-			if err == nil {
-				if _, err := s.execHook(tx,
-					`UPDATE observations
-					 SET type = ?,
-					     title = ?,
-					     content = ?,
-					     tool_name = ?,
-					     topic_key = ?,
-					     normalized_hash = ?,
-					     revision_count = revision_count + 1,
-					     last_seen_at = datetime('now'),
-					     updated_at = datetime('now')
-					 WHERE id = ?`,
-					p.Type,
-					title,
-					content,
-					nullableString(p.ToolName),
-					nullableString(topicKey),
-					normHash,
-					existingID,
-				); err != nil {
-					return err
-				}
-				observationID = existingID
-				if len(p.Tags) > 0 {
-					if err := s.setTagsForObservationTx(tx, existingID, p.Tags); err != nil {
-						return err
-					}
-				}
-				obs, err = s.getObservationTx(tx, existingID)
-				if err != nil {
-					return err
-				}
-				return s.enqueueSyncMutationTx(tx, SyncEntityObservation, obs.SyncID, SyncOpUpsert, observationPayloadFromObservation(obs))
-			}
-			if err != sql.ErrNoRows {
-				return err
-			}
-		}
-
-		window := dedupeWindowExpression(s.cfg.DedupeWindow)
-		var existingID int64
-		err := tx.QueryRow(
-			`SELECT id FROM observations
-			 WHERE normalized_hash = ?
-			   AND ifnull(project, '') = ifnull(?, '')
-			   AND scope = ?
-			   AND type = ?
-			   AND title = ?
-			   AND deleted_at IS NULL
-			   AND datetime(created_at) >= datetime('now', ?)
-			 ORDER BY created_at DESC
-			 LIMIT 1`,
-			normHash, nullableString(p.Project), scope, p.Type, title, window,
-		).Scan(&existingID)
-		if err == nil {
-			if _, err := s.execHook(tx,
-				`UPDATE observations
-				 SET duplicate_count = duplicate_count + 1,
-				     last_seen_at = datetime('now'),
-				     updated_at = datetime('now')
-				 WHERE id = ?`,
-				existingID,
-			); err != nil {
-				return err
-			}
-			obs, err = s.getObservationTx(tx, existingID)
-			if err != nil {
-				return err
-			}
-			observationID = existingID
-			return s.enqueueSyncMutationTx(tx, SyncEntityObservation, obs.SyncID, SyncOpUpsert, observationPayloadFromObservation(obs))
-		}
-		if err != sql.ErrNoRows {
-			return err
-		}
-
-		syncID := newSyncID("obs")
-		res, err := s.execHook(tx,
-			`INSERT INTO observations (sync_id, session_id, type, title, content, tool_name, project, scope, topic_key, normalized_hash, revision_count, duplicate_count, last_seen_at, updated_at)
-			 VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 1, 1, datetime('now'), datetime('now'))`,
-			syncID, p.SessionID, p.Type, title, content,
-			nullableString(p.ToolName), nullableString(p.Project), scope, nullableString(topicKey), normHash,
-		)
-		if err != nil {
-			return err
-		}
-		observationID, err = res.LastInsertId()
-		if err != nil {
-			return err
-		}
-		if len(p.Tags) > 0 {
-			if err := s.setTagsForObservationTx(tx, observationID, p.Tags); err != nil {
-				return err
-			}
-		}
-		obs, err = s.getObservationTx(tx, observationID)
-		if err != nil {
-			return err
-		}
-		return s.enqueueSyncMutationTx(tx, SyncEntityObservation, obs.SyncID, SyncOpUpsert, observationPayloadFromObservation(obs))
-	})
-	if err != nil {
-		return 0, err
-	}
-	return observationID, nil
-}
-
-// RecentObservations returns recent observations for context retrieval.
-// Optional tags parameter filters to observations that have ALL specified tags.
-func (s *Store) RecentObservations(project, scope string, limit int, tags ...string) ([]Observation, error) {
-	return s.recentObservations(project, scope, limit, ContextOptions{Tags: tags})
-}
-
-// RecentObservationsOpts is like RecentObservations but accepts full ContextOptions.
-func (s *Store) RecentObservationsOpts(project, scope string, limit int, opts ContextOptions) ([]Observation, error) {
-	return s.recentObservations(project, scope, limit, opts)
-}
-
-func (s *Store) recentObservations(project, scope string, limit int, opts ContextOptions) ([]Observation, error) {
-	if limit <= 0 {
-		limit = s.cfg.MaxContextResults
-	}
-
-	query := `
-		SELECT o.id, ifnull(o.sync_id, '') as sync_id, o.session_id, o.type, o.title, o.content, o.tool_name, o.project,
-		       o.scope, o.topic_key, o.revision_count, o.duplicate_count, o.last_seen_at, o.created_at, o.updated_at, o.deleted_at
-		FROM observations o
-		WHERE o.deleted_at IS NULL`
-	var args []any
-
-	// Hard filters (WHERE clause) — args must come before boost args.
-	if project != "" {
-		query += " AND o.project = ?"
-		args = append(args, project)
-	}
-	if scope != "" {
-		query += " AND o.scope = ?"
-		args = append(args, normalizeScope(scope))
-	}
-	for _, tag := range opts.Tags {
-		if norm := normalizeTag(tag); norm != "" {
-			query += " AND o.id IN (SELECT observation_id FROM observation_tags WHERE tag = ?)"
-			args = append(args, norm)
-		}
-	}
-
-	// Soft boost (ORDER BY expression) — args must come after WHERE args.
-	boostTags := normalizeTagList(opts.PreferTags)
-	topicKey := strings.TrimSpace(opts.TopicKey)
-
-	if topicKey != "" || len(boostTags) > 0 {
-		var parts []string
-		var boostArgs []any
-		if topicKey != "" {
-			parts = append(parts, "(CASE WHEN o.topic_key = ? THEN 1 ELSE 0 END)")
-			boostArgs = append(boostArgs, topicKey)
-		}
-		if len(boostTags) > 0 {
-			placeholders := strings.Repeat("?,", len(boostTags))
-			placeholders = placeholders[:len(placeholders)-1]
-			parts = append(parts, fmt.Sprintf(
-				"(SELECT COUNT(*) FROM observation_tags bt WHERE bt.observation_id = o.id AND bt.tag IN (%s))",
-				placeholders,
-			))
-			for _, t := range boostTags {
-				boostArgs = append(boostArgs, t)
-			}
-		}
-		boostExpr := strings.Join(parts, " + ")
-		query += fmt.Sprintf(" ORDER BY (%s) DESC, o.created_at DESC, o.id DESC LIMIT ?", boostExpr)
-		args = append(args, boostArgs...)
-	} else {
-		query += " ORDER BY o.created_at DESC, o.id DESC LIMIT ?"
-	}
-	args = append(args, limit)
-
-	return s.queryObservations(query, args...)
-}
-
-// ─── User Prompts ────────────────────────────────────────────────────────────
-
-func (s *Store) AddPrompt(p AddPromptParams) (int64, error) {
-	content := stripPrivateTags(p.Content)
-	if len(content) > s.cfg.MaxObservationLength {
-		content = content[:s.cfg.MaxObservationLength] + "... [truncated]"
-	}
-
-	var promptID int64
-	err := s.withTx(func(tx *sql.Tx) error {
-		syncID := newSyncID("prompt")
-		res, err := s.execHook(tx,
-			`INSERT INTO user_prompts (sync_id, session_id, content, project) VALUES (?, ?, ?, ?)`,
-			syncID, p.SessionID, content, nullableString(p.Project),
-		)
-		if err != nil {
-			return err
-		}
-		promptID, err = res.LastInsertId()
-		if err != nil {
-			return err
-		}
-		return s.enqueueSyncMutationTx(tx, SyncEntityPrompt, syncID, SyncOpUpsert, syncPromptPayload{
-			SyncID:    syncID,
-			SessionID: p.SessionID,
-			Content:   content,
-			Project:   nullableString(p.Project),
-		})
-	})
-	if err != nil {
-		return 0, err
-	}
-	return promptID, nil
-}
-
-func (s *Store) RecentPrompts(project string, limit int) ([]Prompt, error) {
-	if limit <= 0 {
-		limit = 20
-	}
-
-	query := `SELECT id, ifnull(sync_id, '') as sync_id, session_id, content, ifnull(project, '') as project, created_at FROM user_prompts`
-	args := []any{}
-
-	if project != "" {
-		query += " WHERE project = ?"
-		args = append(args, project)
-	}
-
-	query += " ORDER BY created_at DESC LIMIT ?"
-	args = append(args, limit)
-
-	rows, err := s.queryItHook(s.db, query, args...)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var results []Prompt
-	for rows.Next() {
-		var p Prompt
-		if err := rows.Scan(&p.ID, &p.SyncID, &p.SessionID, &p.Content, &p.Project, &p.CreatedAt); err != nil {
-			return nil, err
-		}
-		results = append(results, p)
-	}
-	return results, rows.Err()
-}
-
-func (s *Store) SearchPrompts(query string, project string, limit int) ([]Prompt, error) {
-	if limit <= 0 {
-		limit = 10
-	}
-
-	ftsQuery := sanitizeFTS(query)
-
-	sql := `
-		SELECT p.id, ifnull(p.sync_id, '') as sync_id, p.session_id, p.content, ifnull(p.project, '') as project, p.created_at
-		FROM prompts_fts fts
-		JOIN user_prompts p ON p.id = fts.rowid
-		WHERE prompts_fts MATCH ?
-	`
-	args := []any{ftsQuery}
-
-	if project != "" {
-		sql += " AND p.project = ?"
-		args = append(args, project)
-	}
-
-	sql += " ORDER BY fts.rank LIMIT ?"
-	args = append(args, limit)
-
-	rows, err := s.queryItHook(s.db, sql, args...)
-	if err != nil {
-		return nil, fmt.Errorf("search prompts: %w", err)
-	}
-	defer rows.Close()
-
-	var results []Prompt
-	for rows.Next() {
-		var p Prompt
-		if err := rows.Scan(&p.ID, &p.SyncID, &p.SessionID, &p.Content, &p.Project, &p.CreatedAt); err != nil {
-			return nil, err
-		}
-		results = append(results, p)
-	}
-	return results, rows.Err()
-}
-
-// ─── Get Single Observation ──────────────────────────────────────────────────
-
-func (s *Store) GetObservation(id int64) (*Observation, error) {
-	row := s.db.QueryRow(
-		`SELECT id, ifnull(sync_id, '') as sync_id, session_id, type, title, content, tool_name, project,
-		        scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		 FROM observations WHERE id = ? AND deleted_at IS NULL`, id,
-	)
-	var o Observation
-	if err := row.Scan(
-		&o.ID, &o.SyncID, &o.SessionID, &o.Type, &o.Title, &o.Content,
-		&o.ToolName, &o.Project, &o.Scope, &o.TopicKey, &o.RevisionCount, &o.DuplicateCount, &o.LastSeenAt,
-		&o.CreatedAt, &o.UpdatedAt, &o.DeletedAt,
-	); err != nil {
-		return nil, err
-	}
-	s.loadTagsForObservation(&o)
-	return &o, nil
-}
-
-func (s *Store) UpdateObservation(id int64, p UpdateObservationParams) (*Observation, error) {
-	var updated *Observation
-	err := s.withTx(func(tx *sql.Tx) error {
-		obs, err := s.getObservationTx(tx, id)
-		if err != nil {
-			return err
-		}
-
-		typ := obs.Type
-		title := obs.Title
-		content := obs.Content
-		project := derefString(obs.Project)
-		scope := obs.Scope
-		topicKey := derefString(obs.TopicKey)
-
-		if p.Type != nil {
-			typ = *p.Type
-		}
-		if p.Title != nil {
-			title = stripPrivateTags(*p.Title)
-		}
-		if p.Content != nil {
-			content = stripPrivateTags(*p.Content)
-			if len(content) > s.cfg.MaxObservationLength {
-				content = content[:s.cfg.MaxObservationLength] + "... [truncated]"
-			}
-		}
-		if p.Project != nil {
-			project = *p.Project
-		}
-		if p.Scope != nil {
-			scope = normalizeScope(*p.Scope)
-		}
-		if p.TopicKey != nil {
-			topicKey = normalizeTopicKey(*p.TopicKey)
-		}
-
-		if _, err := s.execHook(tx,
-			`UPDATE observations
-			 SET type = ?,
-			     title = ?,
-			     content = ?,
-			     project = ?,
-			     scope = ?,
-			     topic_key = ?,
-			     normalized_hash = ?,
-			     revision_count = revision_count + 1,
-			     updated_at = datetime('now')
-			 WHERE id = ? AND deleted_at IS NULL`,
-			typ,
-			title,
-			content,
-			nullableString(project),
-			scope,
-			nullableString(topicKey),
-			hashNormalized(content),
-			id,
-		); err != nil {
-			return err
-		}
-
-		if p.Tags != nil {
-			if err := s.setTagsForObservationTx(tx, id, *p.Tags); err != nil {
-				return err
-			}
-		}
-		updated, err = s.getObservationTx(tx, id)
-		if err != nil {
-			return err
-		}
-		return s.enqueueSyncMutationTx(tx, SyncEntityObservation, updated.SyncID, SyncOpUpsert, observationPayloadFromObservation(updated))
-	})
-	if err != nil {
-		return nil, err
-	}
-	s.loadTagsForObservation(updated)
-	return updated, nil
-}
-
-func (s *Store) DeleteObservation(id int64, hardDelete bool) error {
-	return s.withTx(func(tx *sql.Tx) error {
-		obs, err := s.getObservationTx(tx, id)
-		if err == sql.ErrNoRows {
-			return nil
-		}
-		if err != nil {
-			return err
-		}
-
-		deletedAt := Now()
-		if hardDelete {
-			if _, err := s.execHook(tx, `DELETE FROM observations WHERE id = ?`, id); err != nil {
-				return err
-			}
-		} else {
-			if _, err := s.execHook(tx,
-				`UPDATE observations
-				 SET deleted_at = datetime('now'),
-				     updated_at = datetime('now')
-				 WHERE id = ? AND deleted_at IS NULL`,
-				id,
-			); err != nil {
-				return err
-			}
-			if err := tx.QueryRow(`SELECT deleted_at FROM observations WHERE id = ?`, id).Scan(&deletedAt); err != nil {
-				return err
-			}
-		}
-
-		return s.enqueueSyncMutationTx(tx, SyncEntityObservation, obs.SyncID, SyncOpDelete, syncObservationPayload{
-			SyncID:     obs.SyncID,
-			Deleted:    true,
-			DeletedAt:  &deletedAt,
-			HardDelete: hardDelete,
-		})
-	})
-}
-
-// ─── Timeline ────────────────────────────────────────────────────────────────
-//
-// Timeline provides chronological context around a specific observation.
-// Given an observation ID, it returns N observations before and M after,
-// all within the same session. This is the "progressive disclosure" pattern
-// from claude-mem — agents first search, then use timeline to drill into
-// the chronological neighborhood of a result.
-
-func (s *Store) Timeline(observationID int64, before, after int) (*TimelineResult, error) {
-	if before <= 0 {
-		before = 5
-	}
-	if after <= 0 {
-		after = 5
-	}
-
-	// 1. Get the focus observation
-	focus, err := s.GetObservation(observationID)
-	if err != nil {
-		return nil, fmt.Errorf("timeline: observation #%d not found: %w", observationID, err)
-	}
-
-	// 2. Get session info
-	session, err := s.GetSession(focus.SessionID)
-	if err != nil {
-		// Session might be missing for manual-save observations — non-fatal
-		session = nil
-	}
-
-	// 3. Get observations BEFORE the focus (same session, older, chronological order)
-	beforeRows, err := s.queryItHook(s.db, `
-		SELECT id, session_id, type, title, content, tool_name, project,
-		       scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		FROM observations
-		WHERE session_id = ? AND id < ? AND deleted_at IS NULL
-		ORDER BY id DESC
-		LIMIT ?
-	`, focus.SessionID, observationID, before)
-	if err != nil {
-		return nil, fmt.Errorf("timeline: before query: %w", err)
-	}
-	defer beforeRows.Close()
-
-	var beforeEntries []TimelineEntry
-	for beforeRows.Next() {
-		var e TimelineEntry
-		if err := beforeRows.Scan(
-			&e.ID, &e.SessionID, &e.Type, &e.Title, &e.Content,
-			&e.ToolName, &e.Project, &e.Scope, &e.TopicKey, &e.RevisionCount, &e.DuplicateCount, &e.LastSeenAt,
-			&e.CreatedAt, &e.UpdatedAt, &e.DeletedAt,
-		); err != nil {
-			return nil, err
-		}
-		beforeEntries = append(beforeEntries, e)
-	}
-	if err := beforeRows.Err(); err != nil {
-		return nil, err
-	}
-	// Reverse to get chronological order (oldest first)
-	for i, j := 0, len(beforeEntries)-1; i < j; i, j = i+1, j-1 {
-		beforeEntries[i], beforeEntries[j] = beforeEntries[j], beforeEntries[i]
-	}
-
-	// 4. Get observations AFTER the focus (same session, newer, chronological order)
-	afterRows, err := s.queryItHook(s.db, `
-		SELECT id, session_id, type, title, content, tool_name, project,
-		       scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		FROM observations
-		WHERE session_id = ? AND id > ? AND deleted_at IS NULL
-		ORDER BY id ASC
-		LIMIT ?
-	`, focus.SessionID, observationID, after)
-	if err != nil {
-		return nil, fmt.Errorf("timeline: after query: %w", err)
-	}
-	defer afterRows.Close()
-
-	var afterEntries []TimelineEntry
-	for afterRows.Next() {
-		var e TimelineEntry
-		if err := afterRows.Scan(
-			&e.ID, &e.SessionID, &e.Type, &e.Title, &e.Content,
-			&e.ToolName, &e.Project, &e.Scope, &e.TopicKey, &e.RevisionCount, &e.DuplicateCount, &e.LastSeenAt,
-			&e.CreatedAt, &e.UpdatedAt, &e.DeletedAt,
-		); err != nil {
-			return nil, err
-		}
-		afterEntries = append(afterEntries, e)
-	}
-	if err := afterRows.Err(); err != nil {
-		return nil, err
-	}
-
-	// 5. Count total observations in the session for context
-	var totalInRange int
-	// COUNT(*) always returns a row; Scan only fails on DB-level errors.
-	_ = s.db.QueryRow(
-		"SELECT COUNT(*) FROM observations WHERE session_id = ? AND deleted_at IS NULL", focus.SessionID,
-	).Scan(&totalInRange)
-
-	return &TimelineResult{
-		Focus:        *focus,
-		Before:       beforeEntries,
-		After:        afterEntries,
-		SessionInfo:  session,
-		TotalInRange: totalInRange,
-	}, nil
-}
-
-// ─── Search (FTS5) ───────────────────────────────────────────────────────────
-
-func (s *Store) Search(query string, opts SearchOptions) ([]SearchResult, error) {
-	limit := opts.Limit
-	if limit <= 0 {
-		limit = 10
-	}
-	if limit > s.cfg.MaxSearchResults {
-		limit = s.cfg.MaxSearchResults
-	}
-
-	if strings.TrimSpace(query) == "" {
-		return s.searchByFilter(opts, limit)
-	}
-	return s.searchFTS(query, opts, limit)
-}
-
-// searchByFilter handles Search when query is empty — no FTS, direct table scan.
-// Orders by PreferTags overlap (desc) then created_at (desc).
-func (s *Store) searchByFilter(opts SearchOptions, limit int) ([]SearchResult, error) {
-	boostTags := normalizeTagList(opts.PreferTags)
-
-	var boostExpr string
-	var boostArgs []any
-	if len(boostTags) > 0 {
-		placeholders := strings.Repeat("?,", len(boostTags))
-		placeholders = placeholders[:len(placeholders)-1]
-		boostExpr = fmt.Sprintf(
-			"(SELECT COUNT(*) FROM observation_tags bt WHERE bt.observation_id = o.id AND bt.tag IN (%s))",
-			placeholders,
-		)
-		for _, t := range boostTags {
-			boostArgs = append(boostArgs, t)
-		}
-	} else {
-		boostExpr = "0"
-	}
-
-	q := fmt.Sprintf(`
-		SELECT o.id, ifnull(o.sync_id, '') as sync_id, o.session_id, o.type, o.title, o.content, o.tool_name, o.project,
-		       o.scope, o.topic_key, o.revision_count, o.duplicate_count, o.last_seen_at, o.created_at, o.updated_at, o.deleted_at,
-		       CAST(%s AS REAL) as rank
-		FROM observations o
-		WHERE o.deleted_at IS NULL`, boostExpr)
-	args := append([]any{}, boostArgs...)
-
-	q, args = applyObservationFilters(q, args, opts)
-	q += " ORDER BY rank DESC, o.created_at DESC, o.id DESC LIMIT ?"
-	args = append(args, limit)
-
-	return s.runSearchQuery(q, args)
-}
-
-// searchFTS handles Search when a text query is provided.
-// Tag hard-filter and TopicKey filter are applied as WHERE clauses.
-// PreferTags add a secondary sort by overlap count after FTS rank.
-func (s *Store) searchFTS(query string, opts SearchOptions, limit int) ([]SearchResult, error) {
-	ftsQuery := sanitizeFTS(query)
-	boostTags := normalizeTagList(opts.PreferTags)
-
-	var boostExpr string
-	var boostArgs []any
-	if len(boostTags) > 0 {
-		placeholders := strings.Repeat("?,", len(boostTags))
-		placeholders = placeholders[:len(placeholders)-1]
-		boostExpr = fmt.Sprintf(
-			"(SELECT COUNT(*) FROM observation_tags bt WHERE bt.observation_id = o.id AND bt.tag IN (%s))",
-			placeholders,
-		)
-		for _, t := range boostTags {
-			boostArgs = append(boostArgs, t)
-		}
-	} else {
-		boostExpr = "0"
-	}
-
-	q := fmt.Sprintf(`
-		SELECT o.id, ifnull(o.sync_id, '') as sync_id, o.session_id, o.type, o.title, o.content, o.tool_name, o.project,
-		       o.scope, o.topic_key, o.revision_count, o.duplicate_count, o.last_seen_at, o.created_at, o.updated_at, o.deleted_at,
-		       fts.rank - CAST(%s AS REAL) * 0.5 as rank
-		FROM observations_fts fts
-		JOIN observations o ON o.id = fts.rowid
-		WHERE observations_fts MATCH ? AND o.deleted_at IS NULL`, boostExpr)
-	args := append([]any{}, boostArgs...)
-	args = append(args, ftsQuery)
-
-	q, args = applyObservationFilters(q, args, opts)
-	q += " ORDER BY rank LIMIT ?"
-	args = append(args, limit)
-
-	return s.runSearchQuery(q, args)
-}
-
-// applyObservationFilters appends WHERE clauses for Type, Project, Scope, TopicKey and Tags.
-func applyObservationFilters(q string, args []any, opts SearchOptions) (string, []any) {
-	if opts.Type != "" {
-		q += " AND o.type = ?"
-		args = append(args, opts.Type)
-	}
-	if opts.Project != "" {
-		q += " AND o.project = ?"
-		args = append(args, opts.Project)
-	}
-	if opts.Scope != "" {
-		q += " AND o.scope = ?"
-		args = append(args, normalizeScope(opts.Scope))
-	}
-	if opts.TopicKey != "" {
-		q += " AND o.topic_key = ?"
-		args = append(args, opts.TopicKey)
-	}
-	for _, tag := range opts.Tags {
-		if norm := normalizeTag(tag); norm != "" {
-			q += " AND o.id IN (SELECT observation_id FROM observation_tags WHERE tag = ?)"
-			args = append(args, norm)
-		}
-	}
-	return q, args
-}
-
-func (s *Store) runSearchQuery(q string, args []any) ([]SearchResult, error) {
-	rows, err := s.queryItHook(s.db, q, args...)
-	if err != nil {
-		return nil, fmt.Errorf("search: %w", err)
-	}
-	defer rows.Close()
-
-	var results []SearchResult
-	for rows.Next() {
-		var sr SearchResult
-		if err := rows.Scan(
-			&sr.ID, &sr.SyncID, &sr.SessionID, &sr.Type, &sr.Title, &sr.Content,
-			&sr.ToolName, &sr.Project, &sr.Scope, &sr.TopicKey, &sr.RevisionCount, &sr.DuplicateCount,
-			&sr.LastSeenAt, &sr.CreatedAt, &sr.UpdatedAt, &sr.DeletedAt,
-			&sr.Rank,
-		); err != nil {
-			return nil, err
-		}
-		results = append(results, sr)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	s.loadTagsForSearchResults(results)
-	return results, nil
-}
-
-// normalizeTagList normalizes and deduplicates a slice of raw tag strings.
-func normalizeTagList(raw []string) []string {
-	seen := make(map[string]bool)
-	var out []string
-	for _, r := range raw {
-		if n := normalizeTag(r); n != "" && !seen[n] {
-			seen[n] = true
-			out = append(out, n)
-		}
-	}
-	return out
-}
-
-// ─── Tags ────────────────────────────────────────────────────────────────────
-
-// ListTags returns the tags in use for a project, ordered by frequency.
-// Includes last_used_at derived from the most recent observation with that tag.
-func (s *Store) ListTags(project string) ([]TagInfo, error) {
-	q := `
-		SELECT ot.tag,
-		       COUNT(*) AS count,
-		       MAX(o.created_at) AS last_used_at
-		FROM observation_tags ot
-		JOIN observations o ON o.id = ot.observation_id
-		WHERE o.deleted_at IS NULL`
-	var args []any
-	if project != "" {
-		q += " AND o.project = ?"
-		args = append(args, project)
-	}
-	q += " GROUP BY ot.tag ORDER BY count DESC, ot.tag ASC"
-
-	rows, err := s.queryItHook(s.db, q, args...)
-	if err != nil {
-		return nil, fmt.Errorf("list tags: %w", err)
-	}
-	defer rows.Close()
-
-	var tags []TagInfo
-	for rows.Next() {
-		var ti TagInfo
-		if err := rows.Scan(&ti.Tag, &ti.Count, &ti.LastUsedAt); err != nil {
-			return nil, err
-		}
-		tags = append(tags, ti)
-	}
-	return tags, rows.Err()
-}
-
-// TagStats returns tags for a project filtered and ordered according to opts.
-// It does not modify ListTags behaviour — both coexist independently.
-func (s *Store) TagStats(project string, opts TagStatsOptions) ([]TagInfo, error) {
-	q := `
-		SELECT ot.tag,
-		       COUNT(*) AS cnt,
-		       MAX(datetime(o.created_at)) AS last_used_at
-		FROM observation_tags ot
-		JOIN observations o ON o.id = ot.observation_id
-		WHERE o.deleted_at IS NULL`
-
-	var args []any
-	if project != "" {
-		q += " AND o.project = ?"
-		args = append(args, project)
-	}
-	q += " GROUP BY ot.tag"
-
-	var having []string
-	if opts.MinCount > 0 {
-		having = append(having, "COUNT(*) >= ?")
-		args = append(args, opts.MinCount)
-	}
-	if opts.MaxCount > 0 {
-		having = append(having, "COUNT(*) <= ?")
-		args = append(args, opts.MaxCount)
-	}
-	if !opts.UnusedSince.IsZero() {
-		having = append(having, "MAX(datetime(o.created_at)) < datetime(?)")
-		args = append(args, opts.UnusedSince.UTC().Format(time.RFC3339))
-	}
-	if len(having) > 0 {
-		q += " HAVING " + strings.Join(having, " AND ")
-	}
-
-	switch opts.SortBy {
-	case "stale":
-		q += " ORDER BY last_used_at ASC, ot.tag ASC"
-	case "alpha":
-		q += " ORDER BY ot.tag ASC"
-	default:
-		q += " ORDER BY cnt DESC, ot.tag ASC"
-	}
-
-	if opts.Limit > 0 {
-		q += " LIMIT ?"
-		args = append(args, opts.Limit)
-	}
-
-	rows, err := s.queryItHook(s.db, q, args...)
-	if err != nil {
-		return nil, fmt.Errorf("tag stats: %w", err)
-	}
-	defer rows.Close()
-
-	var tags []TagInfo
-	for rows.Next() {
-		var ti TagInfo
-		if err := rows.Scan(&ti.Tag, &ti.Count, &ti.LastUsedAt); err != nil {
-			return nil, err
-		}
-		tags = append(tags, ti)
-	}
-	return tags, rows.Err()
-}
-
-// RelatedTags returns tags that co-occur with the given tag in observations and/or sessions.
-// When both IncludeObservations and IncludeSessions are false, both sources are used.
-func (s *Store) RelatedTags(project, tag string, opts RelatedTagsOptions) ([]RelatedTag, error) {
-	tag = NormalizeTag(tag)
-	if tag == "" {
-		return nil, fmt.Errorf("tag must not be empty")
-	}
-
-	// If neither flag is set, include both sources.
-	includeObs := opts.IncludeObservations || (!opts.IncludeObservations && !opts.IncludeSessions)
-	includeSes := opts.IncludeSessions || (!opts.IncludeObservations && !opts.IncludeSessions)
-
-	// acc accumulates counts and last-seen dates per co-occurring tag.
-	type entry struct {
-		count    int
-		lastSeen string
-	}
-	acc := make(map[string]*entry)
-
-	maxDate := func(a, b string) string {
-		if a > b {
-			return a
-		}
-		return b
-	}
-
-	if includeObs {
-		q := `
-			SELECT ot2.tag, COUNT(*) AS cnt, MAX(o.created_at) AS last_seen
-			FROM observation_tags ot1
-			JOIN observation_tags ot2
-			  ON ot1.observation_id = ot2.observation_id AND ot2.tag != ot1.tag
-			JOIN observations o ON o.id = ot1.observation_id
-			WHERE ot1.tag = ?
-			  AND o.deleted_at IS NULL`
-		var args []any
-		args = append(args, tag)
-		if project != "" {
-			q += " AND o.project = ?"
-			args = append(args, project)
-		}
-		if !opts.Since.IsZero() {
-			q += " AND o.created_at >= datetime(?)"
-			args = append(args, opts.Since.UTC().Format(time.RFC3339))
-		}
-		q += " GROUP BY ot2.tag"
-
-		rows, err := s.queryItHook(s.db, q, args...)
-		if err != nil {
-			return nil, fmt.Errorf("related tags (observations): %w", err)
-		}
-		defer rows.Close()
-		for rows.Next() {
-			var t string
-			var cnt int
-			var last string
-			if err := rows.Scan(&t, &cnt, &last); err != nil {
-				return nil, err
-			}
-			if e, ok := acc[t]; ok {
-				e.count += cnt
-				e.lastSeen = maxDate(e.lastSeen, last)
-			} else {
-				acc[t] = &entry{count: cnt, lastSeen: last}
-			}
-		}
-		if err := rows.Err(); err != nil {
-			return nil, err
-		}
-	}
-
-	if includeSes {
-		q := `
-			SELECT st2.tag, COUNT(*) AS cnt, MAX(s.started_at) AS last_seen
-			FROM session_tags st1
-			JOIN session_tags st2
-			  ON st1.session_id = st2.session_id AND st2.tag != st1.tag
-			JOIN sessions s ON s.id = st1.session_id
-			WHERE st1.tag = ?`
-		var args []any
-		args = append(args, tag)
-		if project != "" {
-			q += " AND s.project = ?"
-			args = append(args, project)
-		}
-		if !opts.Since.IsZero() {
-			q += " AND s.started_at >= datetime(?)"
-			args = append(args, opts.Since.UTC().Format(time.RFC3339))
-		}
-		q += " GROUP BY st2.tag"
-
-		rows, err := s.queryItHook(s.db, q, args...)
-		if err != nil {
-			return nil, fmt.Errorf("related tags (sessions): %w", err)
-		}
-		defer rows.Close()
-		for rows.Next() {
-			var t string
-			var cnt int
-			var last string
-			if err := rows.Scan(&t, &cnt, &last); err != nil {
-				return nil, err
-			}
-			if e, ok := acc[t]; ok {
-				e.count += cnt
-				e.lastSeen = maxDate(e.lastSeen, last)
-			} else {
-				acc[t] = &entry{count: cnt, lastSeen: last}
-			}
-		}
-		if err := rows.Err(); err != nil {
-			return nil, err
-		}
-	}
-
-	// Build result slice, apply MinCooccurrence filter.
-	minC := opts.MinCooccurrence
-	if minC <= 0 {
-		minC = 1
-	}
-	var result []RelatedTag
-	for t, e := range acc {
-		if e.count < minC {
-			continue
-		}
-		result = append(result, RelatedTag{
-			Tag:               t,
-			CooccurrenceCount: e.count,
-			Score:             float64(e.count),
-			LastSeenAt:        e.lastSeen,
-		})
-	}
-
-	// Sort: score desc, tag asc for ties.
-	sort.Slice(result, func(i, j int) bool {
-		if result[i].Score != result[j].Score {
-			return result[i].Score > result[j].Score
-		}
-		return result[i].Tag < result[j].Tag
-	})
-
-	if opts.Limit > 0 && len(result) > opts.Limit {
-		result = result[:opts.Limit]
-	}
-	return result, nil
-}
-
-// tagWeights computes observability signals for a slice of TagInfo.
-// topN controls the IsDominant threshold: the top N tags by frequency are marked dominant.
-// RecencyScore is normalised to [0.0, 1.0] within the provided set.
-// Not yet used in retrieval — prepared for a future weighting phase.
-// parseTagTimestamp parses a tag's LastUsedAt field, trying RFC3339 first and
-// then the SQLite default datetime layout as a fallback.
-func parseTagTimestamp(s string) (time.Time, error) {
-	if t, err := time.Parse(time.RFC3339, s); err == nil {
-		return t, nil
-	}
-	return time.ParseInLocation("2006-01-02 15:04:05", s, time.UTC)
-}
-
-func tagWeights(tags []TagInfo, topN int) []TagWeight {
-	if len(tags) == 0 {
-		return nil
-	}
-
-	// Clamp topN to avoid a negative map capacity.
-	n := topN
-	if n < 0 {
-		n = 0
-	}
-
-	// Determine dominant tags by frequency (independent of input order).
-	sorted := make([]TagInfo, len(tags))
-	copy(sorted, tags)
-	sort.Slice(sorted, func(i, j int) bool { return sorted[i].Count > sorted[j].Count })
-	dominant := make(map[string]bool, n)
-	for i, ti := range sorted {
-		if i >= n {
-			break
-		}
-		dominant[ti.Tag] = true
-	}
-
-	// Find timestamp range for RecencyScore normalisation.
-	var oldest, newest time.Time
-	for _, ti := range tags {
-		t, err := parseTagTimestamp(ti.LastUsedAt)
-		if err != nil {
-			continue
-		}
-		if oldest.IsZero() || t.Before(oldest) {
-			oldest = t
-		}
-		if newest.IsZero() || t.After(newest) {
-			newest = t
-		}
-	}
-	span := newest.Sub(oldest).Seconds()
-
-	out := make([]TagWeight, len(tags))
-	for i, ti := range tags {
-		w := TagWeight{
-			Tag:        ti.Tag,
-			Frequency:  ti.Count,
-			IsDominant: dominant[ti.Tag],
-		}
-		t, err := parseTagTimestamp(ti.LastUsedAt)
-		if err == nil {
-			if span > 0 {
-				w.RecencyScore = t.Sub(oldest).Seconds() / span
-			} else {
-				w.RecencyScore = 1.0
-			}
-		}
-		out[i] = w
-	}
-	return out
-}
-
-// MergeTags renames every occurrence of fromTag to toTag across all observations
-// and sessions. fromTag is matched against the raw stored value (no alias
-// resolution), which allows migrating legacy tags (e.g. "database" → "db").
-// toTag goes through full normalization including alias resolution.
-// Returns the number of observations and sessions updated.
-func (s *Store) MergeTags(fromTag, toTag string) (obsCount int, sessCount int, err error) {
-	// from uses base normalization only — no alias — to match what may be stored.
-	from := normalizeTagBase(fromTag)
-	to := normalizeTag(toTag)
-	if from == "" || to == "" {
-		return 0, 0, fmt.Errorf("both from and to tags must be non-empty after normalization")
-	}
-	if from == to {
-		return 0, 0, fmt.Errorf("from and to tags are identical after normalization: %q", to)
-	}
-	if isBlockedTag(to) {
-		return 0, 0, fmt.Errorf("target tag %q is blocked (too generic)", to)
-	}
-
-	err = s.withTx(func(tx *sql.Tx) error {
-		// Collect affected observations before modifying tags.
-		obsRows, err := s.queryItHook(tx,
-			`SELECT o.id, o.sync_id, o.session_id, o.type, o.title, o.content,
-			        o.tool_name, o.project, o.scope, o.topic_key
-			 FROM observations o
-			 JOIN observation_tags ot ON ot.observation_id = o.id
-			 WHERE ot.tag = ? AND o.deleted_at IS NULL`, from)
-		if err != nil {
-			return fmt.Errorf("merge: collect observations: %w", err)
-		}
-		type affectedObs struct {
-			id      int64
-			payload syncObservationPayload
-		}
-		var obsList []affectedObs
-		for obsRows.Next() {
-			var r affectedObs
-			if err := obsRows.Scan(&r.id, &r.payload.SyncID, &r.payload.SessionID,
-				&r.payload.Type, &r.payload.Title, &r.payload.Content,
-				&r.payload.ToolName, &r.payload.Project, &r.payload.Scope, &r.payload.TopicKey,
-			); err != nil {
-				obsRows.Close()
-				return fmt.Errorf("merge: scan observation: %w", err)
-			}
-			obsList = append(obsList, r)
-		}
-		obsRows.Close()
-		if err := obsRows.Err(); err != nil {
-			return fmt.Errorf("merge: iterate observations: %w", err)
-		}
-
-		// Collect affected sessions before modifying tags.
-		sessRows, err := s.queryItHook(tx,
-			`SELECT s.id, s.project, s.directory, s.ended_at, s.summary
-			 FROM sessions s
-			 JOIN session_tags st ON st.session_id = s.id
-			 WHERE st.tag = ?`, from)
-		if err != nil {
-			return fmt.Errorf("merge: collect sessions: %w", err)
-		}
-		type affectedSess struct {
-			payload syncSessionPayload
-		}
-		var sessList []affectedSess
-		for sessRows.Next() {
-			var r affectedSess
-			if err := sessRows.Scan(&r.payload.ID, &r.payload.Project,
-				&r.payload.Directory, &r.payload.EndedAt, &r.payload.Summary,
-			); err != nil {
-				sessRows.Close()
-				return fmt.Errorf("merge: scan session: %w", err)
-			}
-			sessList = append(sessList, r)
-		}
-		sessRows.Close()
-		if err := sessRows.Err(); err != nil {
-			return fmt.Errorf("merge: iterate sessions: %w", err)
-		}
-
-		// Promote fromTag → toTag in observations (dedup via INSERT OR IGNORE).
-		if _, err := s.execHook(tx,
-			`INSERT OR IGNORE INTO observation_tags (observation_id, tag)
-			 SELECT observation_id, ? FROM observation_tags WHERE tag = ?`,
-			to, from,
-		); err != nil {
-			return fmt.Errorf("merge observation_tags insert: %w", err)
-		}
-		if _, err := s.execHook(tx, `DELETE FROM observation_tags WHERE tag = ?`, from); err != nil {
-			return fmt.Errorf("merge observation_tags delete: %w", err)
-		}
-		obsCount = len(obsList)
-
-		// Promote fromTag → toTag in sessions.
-		if _, err := s.execHook(tx,
-			`INSERT OR IGNORE INTO session_tags (session_id, tag)
-			 SELECT session_id, ? FROM session_tags WHERE tag = ?`,
-			to, from,
-		); err != nil {
-			return fmt.Errorf("merge session_tags insert: %w", err)
-		}
-		if _, err := s.execHook(tx, `DELETE FROM session_tags WHERE tag = ?`, from); err != nil {
-			return fmt.Errorf("merge session_tags delete: %w", err)
-		}
-		sessCount = len(sessList)
-
-		// Enqueue sync mutations for affected observations.
-		for _, r := range obsList {
-			var o Observation
-			o.ID = r.id
-			s.loadTagsForObservationTx(tx, &o)
-			tags := o.Tags
-			if tags == nil {
-				tags = []string{}
-			}
-			r.payload.Tags = &tags
-			if err := s.enqueueSyncMutationTx(tx, SyncEntityObservation, r.payload.SyncID, SyncOpUpsert, r.payload); err != nil {
-				return fmt.Errorf("merge: enqueue observation sync: %w", err)
-			}
-		}
-
-		// Enqueue sync mutations for affected sessions.
-		for _, r := range sessList {
-			var sess Session
-			sess.ID = r.payload.ID
-			s.loadTagsForSessionTx(tx, &sess)
-			tags := sess.Tags
-			if tags == nil {
-				tags = []string{}
-			}
-			r.payload.Tags = &tags
-			if err := s.enqueueSyncMutationTx(tx, SyncEntitySession, r.payload.ID, SyncOpUpsert, r.payload); err != nil {
-				return fmt.Errorf("merge: enqueue session sync: %w", err)
-			}
-		}
-
-		return nil
-	})
-	return obsCount, sessCount, err
-}
-
-// ─── Stats ───────────────────────────────────────────────────────────────────
-
-func (s *Store) Stats() (*Stats, error) {
-	stats := &Stats{}
-
-	// COUNT(*) always returns a row; Scan only fails on DB-level errors.
-	_ = s.db.QueryRow("SELECT COUNT(*) FROM sessions").Scan(&stats.TotalSessions)
-	_ = s.db.QueryRow("SELECT COUNT(*) FROM observations WHERE deleted_at IS NULL").Scan(&stats.TotalObservations)
-	_ = s.db.QueryRow("SELECT COUNT(*) FROM user_prompts").Scan(&stats.TotalPrompts)
-
-	rows, err := s.queryItHook(s.db, "SELECT project FROM observations WHERE project IS NOT NULL AND deleted_at IS NULL GROUP BY project ORDER BY MAX(created_at) DESC")
-	if err != nil {
-		return stats, nil
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var p string
-		if err := rows.Scan(&p); err == nil {
-			stats.Projects = append(stats.Projects, p)
-		}
-	}
-
-	return stats, nil
-}
-
-// ─── Context Formatting ─────────────────────────────────────────────────────
-
-// FormatContext returns a formatted context string for agent consumption.
-// Optional tags parameter filters the observations included in the context.
-func (s *Store) FormatContext(project, scope string, tags ...string) (string, error) {
-	return s.FormatContextOpts(project, scope, ContextOptions{Tags: tags})
-}
-
-// FormatContextOpts is like FormatContext but accepts full ContextOptions.
-func (s *Store) FormatContextOpts(project, scope string, opts ContextOptions) (string, error) {
-	sessions, err := s.RecentSessions(project, 5)
-	if err != nil {
-		return "", err
-	}
-
-	observations, err := s.recentObservations(project, scope, s.cfg.MaxContextResults, opts)
-	if err != nil {
-		return "", err
-	}
-
-	prompts, err := s.RecentPrompts(project, 10)
-	if err != nil {
-		return "", err
-	}
-
-	if len(sessions) == 0 && len(observations) == 0 && len(prompts) == 0 {
-		return "", nil
-	}
-
-	var b strings.Builder
-	b.WriteString("## Memory from Previous Sessions\n\n")
-
-	if len(sessions) > 0 {
-		b.WriteString("### Recent Sessions\n")
-		for _, sess := range sessions {
-			summary := ""
-			if sess.Summary != nil {
-				summary = fmt.Sprintf(": %s", truncate(*sess.Summary, 200))
-			}
-			fmt.Fprintf(&b, "- **%s** (%s)%s [%d observations]\n",
-				sess.Project, sess.StartedAt, summary, sess.ObservationCount)
-		}
-		b.WriteString("\n")
-	}
-
-	if len(prompts) > 0 {
-		b.WriteString("### Recent User Prompts\n")
-		for _, p := range prompts {
-			fmt.Fprintf(&b, "- %s: %s\n", p.CreatedAt, truncate(p.Content, 200))
-		}
-		b.WriteString("\n")
-	}
-
-	if len(observations) > 0 {
-		b.WriteString("### Recent Observations\n")
-		for _, obs := range observations {
-			fmt.Fprintf(&b, "- [%s] **%s**: %s\n",
-				obs.Type, obs.Title, truncate(obs.Content, 300))
-		}
-		b.WriteString("\n")
-	}
-
-	return b.String(), nil
-}
-
-// ─── Export / Import ─────────────────────────────────────────────────────────
-
-func (s *Store) Export() (*ExportData, error) {
-	data := &ExportData{
-		Version:    "0.1.0",
-		ExportedAt: Now(),
-	}
-
-	// Sessions
-	rows, err := s.queryItHook(s.db,
-		"SELECT id, project, directory, started_at, ended_at, summary FROM sessions ORDER BY started_at",
-	)
-	if err != nil {
-		return nil, fmt.Errorf("export sessions: %w", err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var sess Session
-		if err := rows.Scan(&sess.ID, &sess.Project, &sess.Directory, &sess.StartedAt, &sess.EndedAt, &sess.Summary); err != nil {
-			return nil, err
-		}
-		s.loadTagsForSession(&sess)
-		data.Sessions = append(data.Sessions, sess)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-
-	// Observations
-	obsRows, err := s.queryItHook(s.db,
-		`SELECT id, ifnull(sync_id, '') as sync_id, session_id, type, title, content, tool_name, project,
-		        scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		 FROM observations ORDER BY id`,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("export observations: %w", err)
-	}
-	defer obsRows.Close()
-	for obsRows.Next() {
-		var o Observation
-		if err := obsRows.Scan(
-			&o.ID, &o.SyncID, &o.SessionID, &o.Type, &o.Title, &o.Content,
-			&o.ToolName, &o.Project, &o.Scope, &o.TopicKey, &o.RevisionCount, &o.DuplicateCount, &o.LastSeenAt,
-			&o.CreatedAt, &o.UpdatedAt, &o.DeletedAt,
-		); err != nil {
-			return nil, err
-		}
-		data.Observations = append(data.Observations, o)
-	}
-	if err := obsRows.Err(); err != nil {
-		return nil, err
-	}
-	s.loadTagsForObservations(data.Observations)
-
-	// Prompts
-	promptRows, err := s.queryItHook(s.db,
-		"SELECT id, ifnull(sync_id, '') as sync_id, session_id, content, ifnull(project, '') as project, created_at FROM user_prompts ORDER BY id",
-	)
-	if err != nil {
-		return nil, fmt.Errorf("export prompts: %w", err)
-	}
-	defer promptRows.Close()
-	for promptRows.Next() {
-		var p Prompt
-		if err := promptRows.Scan(&p.ID, &p.SyncID, &p.SessionID, &p.Content, &p.Project, &p.CreatedAt); err != nil {
-			return nil, err
-		}
-		data.Prompts = append(data.Prompts, p)
-	}
-	if err := promptRows.Err(); err != nil {
-		return nil, err
-	}
-
-	return data, nil
-}
-
-func (s *Store) Import(data *ExportData) (*ImportResult, error) {
-	tx, err := s.beginTxHook()
-	if err != nil {
-		return nil, fmt.Errorf("import: begin tx: %w", err)
-	}
-	defer tx.Rollback()
-
-	result := &ImportResult{}
-
-	// Import sessions (skip duplicates)
-	for _, sess := range data.Sessions {
-		res, err := s.execHook(tx,
-			`INSERT OR IGNORE INTO sessions (id, project, directory, started_at, ended_at, summary)
-			 VALUES (?, ?, ?, ?, ?, ?)`,
-			sess.ID, sess.Project, sess.Directory, sess.StartedAt, sess.EndedAt, sess.Summary,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("import session %s: %w", sess.ID, err)
-		}
-		n, _ := res.RowsAffected()
-		if n > 0 && sess.Tags != nil {
-			if err := s.setTagsForSessionTx(tx, sess.ID, sess.Tags); err != nil {
-				return nil, fmt.Errorf("import session %s: tags: %w", sess.ID, err)
-			}
-		}
-		result.SessionsImported += int(n)
-	}
-
-	// Import observations — skip duplicates by normalized_hash (content hash)
-	for _, obs := range data.Observations {
-		hash := hashNormalized(obs.Content)
-		var exists int
-		if err := tx.QueryRow(
-			`SELECT COUNT(*) FROM observations WHERE normalized_hash = ? AND deleted_at IS NULL`, hash,
-		).Scan(&exists); err != nil {
-			return nil, fmt.Errorf("import observation check %d: %w", obs.ID, err)
-		}
-		if exists > 0 {
-			continue
-		}
-		res, err := s.execHook(tx,
-			`INSERT INTO observations (sync_id, session_id, type, title, content, tool_name, project, scope, topic_key, normalized_hash, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at)
-			 VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-			normalizeExistingSyncID(obs.SyncID, "obs"),
-			obs.SessionID,
-			obs.Type,
-			obs.Title,
-			obs.Content,
-			obs.ToolName,
-			obs.Project,
-			normalizeScope(obs.Scope),
-			nullableString(normalizeTopicKey(derefString(obs.TopicKey))),
-			hash,
-			maxInt(obs.RevisionCount, 1),
-			maxInt(obs.DuplicateCount, 1),
-			obs.LastSeenAt,
-			obs.CreatedAt,
-			obs.UpdatedAt,
-			obs.DeletedAt,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("import observation %d: %w", obs.ID, err)
-		}
-		if obs.Tags != nil {
-			newID, err := res.LastInsertId()
-			if err != nil {
-				return nil, fmt.Errorf("import observation %d: last insert id: %w", obs.ID, err)
-			}
-			if err := s.setTagsForObservationTx(tx, newID, obs.Tags); err != nil {
-				return nil, fmt.Errorf("import observation %d: tags: %w", obs.ID, err)
-			}
-		}
-		result.ObservationsImported++
-	}
-
-	// Import prompts
-	for _, p := range data.Prompts {
-		_, err := s.execHook(tx,
-			`INSERT INTO user_prompts (sync_id, session_id, content, project, created_at)
-			 VALUES (?, ?, ?, ?, ?)`,
-			normalizeExistingSyncID(p.SyncID, "prompt"), p.SessionID, p.Content, p.Project, p.CreatedAt,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("import prompt %d: %w", p.ID, err)
-		}
-		result.PromptsImported++
-	}
-
-	if err := s.commitHook(tx); err != nil {
-		return nil, fmt.Errorf("import: commit: %w", err)
-	}
-
-	return result, nil
-}
-
-type ImportResult struct {
-	SessionsImported     int `json:"sessions_imported"`
-	ObservationsImported int `json:"observations_imported"`
-	PromptsImported      int `json:"prompts_imported"`
-}
-
-// ─── Sync Chunk Tracking ─────────────────────────────────────────────────────
-
-// GetSyncedChunks returns a set of chunk IDs that have been imported/exported.
-func (s *Store) GetSyncedChunks() (map[string]bool, error) {
-	rows, err := s.queryItHook(s.db, "SELECT chunk_id FROM sync_chunks")
-	if err != nil {
-		return nil, fmt.Errorf("get synced chunks: %w", err)
-	}
-	defer rows.Close()
-
-	chunks := make(map[string]bool)
-	for rows.Next() {
-		var id string
-		if err := rows.Scan(&id); err != nil {
-			return nil, err
-		}
-		chunks[id] = true
-	}
-	return chunks, rows.Err()
-}
-
-// RecordSyncedChunk marks a chunk as imported/exported so it won't be processed again.
-func (s *Store) RecordSyncedChunk(chunkID string) error {
-	_, err := s.execHook(s.db,
-		"INSERT OR IGNORE INTO sync_chunks (chunk_id) VALUES (?)",
-		chunkID,
-	)
-	return err
-}
-
-// ─── Local Sync State & Mutation Journal ─────────────────────────────────────
-
-func (s *Store) GetSyncState(targetKey string) (*SyncState, error) {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	if err := s.ensureSyncState(targetKey); err != nil {
-		return nil, err
-	}
-	return s.getSyncState(targetKey)
-}
-
-func (s *Store) ListPendingSyncMutations(targetKey string, limit int) ([]SyncMutation, error) {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	if limit <= 0 {
-		limit = 100
-	}
-	// Only return mutations for enrolled projects or empty-project (global) mutations.
-	// Empty-project mutations always sync regardless of enrollment.
-	rows, err := s.queryItHook(s.db, `
-		SELECT sm.seq, sm.target_key, sm.entity, sm.entity_key, sm.op, sm.payload, sm.source, sm.project, sm.occurred_at, sm.acked_at
-		FROM sync_mutations sm
-		LEFT JOIN sync_enrolled_projects sep ON sm.project = sep.project
-		WHERE sm.target_key = ? AND sm.acked_at IS NULL
-		  AND (sm.project = '' OR sep.project IS NOT NULL)
-		ORDER BY sm.seq ASC
-		LIMIT ?`, targetKey, limit)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var mutations []SyncMutation
-	for rows.Next() {
-		var mutation SyncMutation
-		if err := rows.Scan(&mutation.Seq, &mutation.TargetKey, &mutation.Entity, &mutation.EntityKey, &mutation.Op, &mutation.Payload, &mutation.Source, &mutation.Project, &mutation.OccurredAt, &mutation.AckedAt); err != nil {
-			return nil, err
-		}
-		mutations = append(mutations, mutation)
-	}
-	return mutations, rows.Err()
-}
-
-// SkipAckNonEnrolledMutations acks (marks as skipped) all pending mutations
-// that belong to non-enrolled projects, preventing journal bloat. Empty-project
-// mutations are never skipped — they always sync regardless of enrollment.
-func (s *Store) SkipAckNonEnrolledMutations(targetKey string) (int64, error) {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	res, err := s.execHook(s.db, `
-		UPDATE sync_mutations
-		SET acked_at = datetime('now')
-		WHERE target_key = ?
-		  AND acked_at IS NULL
-		  AND project != ''
-		  AND project NOT IN (SELECT project FROM sync_enrolled_projects)`,
-		targetKey,
-	)
-	if err != nil {
-		return 0, err
-	}
-	return res.RowsAffected()
-}
-
-func (s *Store) AckSyncMutations(targetKey string, lastAckedSeq int64) error {
-	if lastAckedSeq <= 0 {
-		return nil
-	}
-	targetKey = normalizeSyncTargetKey(targetKey)
-	return s.withTx(func(tx *sql.Tx) error {
-		state, err := s.getSyncStateTx(tx, targetKey)
-		if err != nil {
-			return err
-		}
-		if _, err := s.execHook(tx,
-			`UPDATE sync_mutations SET acked_at = datetime('now') WHERE target_key = ? AND seq <= ? AND acked_at IS NULL`,
-			targetKey, lastAckedSeq,
-		); err != nil {
-			return err
-		}
-		acked := state.LastAckedSeq
-		if lastAckedSeq > acked {
-			acked = lastAckedSeq
-		}
-		lifecycle := SyncLifecyclePending
-		if acked >= state.LastEnqueuedSeq {
-			lifecycle = SyncLifecycleHealthy
-		}
-		_, err = s.execHook(tx,
-			`UPDATE sync_state
-			 SET last_acked_seq = ?, lifecycle = ?, updated_at = datetime('now')
-			 WHERE target_key = ?`,
-			acked, lifecycle, targetKey,
-		)
-		return err
-	})
-}
-
-// AckSyncMutationSeqs acknowledges specific mutation sequence numbers without
-// requiring them to be contiguous.
-func (s *Store) AckSyncMutationSeqs(targetKey string, seqs []int64) error {
-	if len(seqs) == 0 {
-		return nil
-	}
-	targetKey = normalizeSyncTargetKey(targetKey)
-	return s.withTx(func(tx *sql.Tx) error {
-		state, err := s.getSyncStateTx(tx, targetKey)
-		if err != nil {
-			return err
-		}
-		maxSeq := state.LastAckedSeq
-		for _, seq := range seqs {
-			if seq <= 0 {
-				continue
-			}
-			if _, err := s.execHook(tx,
-				`UPDATE sync_mutations SET acked_at = datetime('now') WHERE target_key = ? AND seq = ? AND acked_at IS NULL`,
-				targetKey, seq,
-			); err != nil {
-				return err
-			}
-			if seq > maxSeq {
-				maxSeq = seq
-			}
-		}
-		var remaining int
-		if err := tx.QueryRow(`SELECT COUNT(*) FROM sync_mutations WHERE target_key = ? AND acked_at IS NULL`, targetKey).Scan(&remaining); err != nil {
-			return err
-		}
-		lifecycle := SyncLifecyclePending
-		if remaining == 0 {
-			lifecycle = SyncLifecycleHealthy
-		}
-		_, err = s.execHook(tx,
-			`UPDATE sync_state SET last_acked_seq = ?, lifecycle = ?, updated_at = datetime('now') WHERE target_key = ?`,
-			maxSeq, lifecycle, targetKey,
-		)
-		return err
-	})
-}
-
-func (s *Store) AcquireSyncLease(targetKey, owner string, ttl time.Duration, now time.Time) (bool, error) {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	if ttl <= 0 {
-		ttl = time.Minute
-	}
-	if now.IsZero() {
-		now = time.Now().UTC()
-	}
-
-	var acquired bool
-	err := s.withTx(func(tx *sql.Tx) error {
-		state, err := s.getSyncStateTx(tx, targetKey)
-		if err != nil {
-			return err
-		}
-		if state.LeaseUntil != nil {
-			leaseUntil, err := time.Parse(time.RFC3339, *state.LeaseUntil)
-			if err == nil && leaseUntil.After(now) && derefString(state.LeaseOwner) != "" && derefString(state.LeaseOwner) != owner {
-				acquired = false
-				return nil
-			}
-		}
-		leaseUntil := now.Add(ttl).UTC().Format(time.RFC3339)
-		_, err = s.execHook(tx,
-			`UPDATE sync_state
-			 SET lease_owner = ?, lease_until = ?, updated_at = datetime('now')
-			 WHERE target_key = ?`,
-			owner, leaseUntil, targetKey,
-		)
-		if err == nil {
-			acquired = true
-		}
-		return err
-	})
-	return acquired, err
-}
-
-func (s *Store) ReleaseSyncLease(targetKey, owner string) error {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	_, err := s.execHook(s.db,
-		`UPDATE sync_state
-		 SET lease_owner = NULL, lease_until = NULL, updated_at = datetime('now')
-		 WHERE target_key = ? AND (lease_owner = ? OR lease_owner IS NULL OR lease_owner = '')`,
-		targetKey, owner,
-	)
-	return err
-}
-
-func (s *Store) MarkSyncFailure(targetKey, message string, backoffUntil time.Time) error {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	backoff := backoffUntil.UTC().Format(time.RFC3339)
-	return s.withTx(func(tx *sql.Tx) error {
-		state, err := s.getSyncStateTx(tx, targetKey)
-		if err != nil {
-			return err
-		}
-		_, err = s.execHook(tx,
-			`UPDATE sync_state
-			 SET lifecycle = ?, consecutive_failures = ?, backoff_until = ?, last_error = ?, updated_at = datetime('now')
-			 WHERE target_key = ?`,
-			SyncLifecycleDegraded, state.ConsecutiveFailures+1, backoff, message, targetKey,
-		)
-		return err
-	})
-}
-
-func (s *Store) MarkSyncHealthy(targetKey string) error {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	_, err := s.execHook(s.db,
-		`UPDATE sync_state
-		 SET lifecycle = ?, consecutive_failures = 0, backoff_until = NULL, last_error = NULL, updated_at = datetime('now')
-		 WHERE target_key = ?`,
-		SyncLifecycleHealthy, targetKey,
-	)
-	return err
-}
-
-func (s *Store) ApplyPulledMutation(targetKey string, mutation SyncMutation) error {
-	targetKey = normalizeSyncTargetKey(targetKey)
-	return s.withTx(func(tx *sql.Tx) error {
-		state, err := s.getSyncStateTx(tx, targetKey)
-		if err != nil {
-			return err
-		}
-		if mutation.Seq <= state.LastPulledSeq {
-			return nil
-		}
-
-		switch mutation.Entity {
-		case SyncEntitySession:
-			var payload syncSessionPayload
-			if err := decodeSyncPayload([]byte(mutation.Payload), &payload); err != nil {
-				return err
-			}
-			if err := s.applySessionPayloadTx(tx, payload); err != nil {
-				return err
-			}
-		case SyncEntityObservation:
-			var payload syncObservationPayload
-			if err := decodeSyncPayload([]byte(mutation.Payload), &payload); err != nil {
-				return err
-			}
-			if mutation.Op == SyncOpDelete {
-				if err := s.applyObservationDeleteTx(tx, payload); err != nil {
-					return err
-				}
-			} else {
-				if err := s.applyObservationUpsertTx(tx, payload); err != nil {
-					return err
-				}
-			}
-		case SyncEntityPrompt:
-			var payload syncPromptPayload
-			if err := decodeSyncPayload([]byte(mutation.Payload), &payload); err != nil {
-				return err
-			}
-			if err := s.applyPromptUpsertTx(tx, payload); err != nil {
-				return err
-			}
-		default:
-			return fmt.Errorf("unknown sync entity %q", mutation.Entity)
-		}
-
-		_, err = s.execHook(tx,
-			`UPDATE sync_state
-			 SET last_pulled_seq = ?, lifecycle = ?, consecutive_failures = 0, backoff_until = NULL, last_error = NULL, updated_at = datetime('now')
-			 WHERE target_key = ?`,
-			mutation.Seq, SyncLifecycleHealthy, targetKey,
-		)
-		return err
-	})
-}
-
-func (s *Store) GetObservationBySyncID(syncID string) (*Observation, error) {
-	row := s.db.QueryRow(
-		`SELECT id, ifnull(sync_id, '') as sync_id, session_id, type, title, content, tool_name, project,
-		        scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		 FROM observations WHERE sync_id = ? AND deleted_at IS NULL ORDER BY id DESC LIMIT 1`,
-		syncID,
-	)
-	var o Observation
-	if err := row.Scan(&o.ID, &o.SyncID, &o.SessionID, &o.Type, &o.Title, &o.Content, &o.ToolName, &o.Project, &o.Scope, &o.TopicKey, &o.RevisionCount, &o.DuplicateCount, &o.LastSeenAt, &o.CreatedAt, &o.UpdatedAt, &o.DeletedAt); err != nil {
-		return nil, err
-	}
-	return &o, nil
-}
-
-// ─── Project Enrollment for Cloud Sync ───────────────────────────────────────
-
-// EnrollProject registers a project for cloud sync. Idempotent — re-enrolling
-// an already-enrolled project is a no-op.
-func (s *Store) EnrollProject(project string) error {
-	if project == "" {
-		return fmt.Errorf("project name must not be empty")
-	}
-	return s.withTx(func(tx *sql.Tx) error {
-		res, err := s.execHook(tx,
-			`INSERT OR IGNORE INTO sync_enrolled_projects (project) VALUES (?)`,
-			project,
-		)
-		if err != nil {
-			return err
-		}
-		rowsAffected, err := res.RowsAffected()
-		if err != nil {
-			return err
-		}
-		if rowsAffected == 0 {
-			return nil
-		}
-		return s.backfillProjectSyncMutationsTx(tx, project)
-	})
-}
-
-// UnenrollProject removes a project from cloud sync enrollment. Idempotent —
-// unenrolling a non-enrolled project is a no-op.
-func (s *Store) UnenrollProject(project string) error {
-	if project == "" {
-		return fmt.Errorf("project name must not be empty")
-	}
-	_, err := s.execHook(s.db,
-		`DELETE FROM sync_enrolled_projects WHERE project = ?`,
-		project,
-	)
-	return err
-}
-
-// ListEnrolledProjects returns all projects currently enrolled for cloud sync,
-// ordered alphabetically by project name.
-func (s *Store) ListEnrolledProjects() ([]EnrolledProject, error) {
-	rows, err := s.queryItHook(s.db,
-		`SELECT project, enrolled_at FROM sync_enrolled_projects ORDER BY project ASC`)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var projects []EnrolledProject
-	for rows.Next() {
-		var ep EnrolledProject
-		if err := rows.Scan(&ep.Project, &ep.EnrolledAt); err != nil {
-			return nil, err
-		}
-		projects = append(projects, ep)
-	}
-	return projects, rows.Err()
-}
-
-// IsProjectEnrolled returns true if the given project is enrolled for cloud sync.
-func (s *Store) IsProjectEnrolled(project string) (bool, error) {
-	var exists int
-	err := s.db.QueryRow(
-		`SELECT 1 FROM sync_enrolled_projects WHERE project = ? LIMIT 1`,
-		project,
-	).Scan(&exists)
-	if err == sql.ErrNoRows {
-		return false, nil
-	}
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-// ─── Project Migration ───────────────────────────────────────────────────────
-
-type MigrateResult struct {
-	Migrated             bool  `json:"migrated"`
-	ObservationsUpdated  int64 `json:"observations_updated"`
-	SessionsUpdated      int64 `json:"sessions_updated"`
-	PromptsUpdated       int64 `json:"prompts_updated"`
-	SyncMutationsUpdated int64 `json:"sync_mutations_updated"`
-}
-
-func (s *Store) MigrateProject(oldName, newName string) (*MigrateResult, error) {
-	if oldName == "" || newName == "" || oldName == newName {
-		return &MigrateResult{}, nil
-	}
-
-	// Check if old project has any records (short-circuit on first match)
-	var exists bool
-	err := s.db.QueryRow(
-		`SELECT EXISTS(
-			SELECT 1 FROM observations WHERE project = ?
-			UNION ALL
-			SELECT 1 FROM sessions WHERE project = ?
-			UNION ALL
-			SELECT 1 FROM user_prompts WHERE project = ?
-			UNION ALL
-			SELECT 1 FROM sync_mutations WHERE project = ?
-			UNION ALL
-			SELECT 1 FROM sync_enrolled_projects WHERE project = ?
-		)`, oldName, oldName, oldName, oldName, oldName,
-	).Scan(&exists)
-	if err != nil {
-		return nil, fmt.Errorf("check old project: %w", err)
-	}
-	if !exists {
-		return &MigrateResult{}, nil
-	}
-
-	result := &MigrateResult{Migrated: true}
-
-	err = s.withTx(func(tx *sql.Tx) error {
-		// FTS triggers handle index updates automatically on UPDATE
-		res, err := s.execHook(tx, `UPDATE observations SET project = ? WHERE project = ?`, newName, oldName)
-		if err != nil {
-			return fmt.Errorf("migrate observations: %w", err)
-		}
-		result.ObservationsUpdated, _ = res.RowsAffected()
-
-		res, err = s.execHook(tx, `UPDATE sessions SET project = ? WHERE project = ?`, newName, oldName)
-		if err != nil {
-			return fmt.Errorf("migrate sessions: %w", err)
-		}
-		result.SessionsUpdated, _ = res.RowsAffected()
-
-		res, err = s.execHook(tx, `UPDATE user_prompts SET project = ? WHERE project = ?`, newName, oldName)
-		if err != nil {
-			return fmt.Errorf("migrate prompts: %w", err)
-		}
-		result.PromptsUpdated, _ = res.RowsAffected()
-
-		res, err = s.execHook(tx,
-			`UPDATE sync_mutations SET project = ?, payload = json_set(payload, '$.project', ?) WHERE project = ?`,
-			newName, newName, oldName)
-		if err != nil {
-			return fmt.Errorf("migrate sync_mutations: %w", err)
-		}
-		result.SyncMutationsUpdated, _ = res.RowsAffected()
-
-		if _, err = s.execHook(tx,
-			`INSERT OR IGNORE INTO sync_enrolled_projects (project, enrolled_at)
-			 SELECT ?, enrolled_at FROM sync_enrolled_projects WHERE project = ?`,
-			newName, oldName); err != nil {
-			return fmt.Errorf("migrate sync_enrolled_projects insert: %w", err)
-		}
-		if _, err = s.execHook(tx,
-			`DELETE FROM sync_enrolled_projects WHERE project = ?`, oldName); err != nil {
-			return fmt.Errorf("migrate sync_enrolled_projects delete: %w", err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-// ─── Helpers ─────────────────────────────────────────────────────────────────
-
-func (s *Store) withTx(fn func(tx *sql.Tx) error) error {
-	tx, err := s.beginTxHook()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-	if err := fn(tx); err != nil {
-		return err
-	}
-	return s.commitHook(tx)
-}
-
-func (s *Store) createSessionTx(tx *sql.Tx, id, project, directory string) error {
-	_, err := s.execHook(tx,
-		`INSERT INTO sessions (id, project, directory) VALUES (?, ?, ?)
-		 ON CONFLICT(id) DO UPDATE SET
-		   project   = CASE WHEN sessions.project = '' THEN excluded.project ELSE sessions.project END,
-		   directory = CASE WHEN sessions.directory = '' THEN excluded.directory ELSE sessions.directory END`,
-		id, project, directory,
-	)
-	return err
-}
-
-func (s *Store) ensureSyncState(targetKey string) error {
-	_, err := s.execHook(s.db,
-		`INSERT OR IGNORE INTO sync_state (target_key, lifecycle, updated_at) VALUES (?, ?, datetime('now'))`,
-		targetKey, SyncLifecycleIdle,
-	)
-	return err
-}
-
-func (s *Store) getSyncState(targetKey string) (*SyncState, error) {
-	row := s.db.QueryRow(`
-		SELECT target_key, lifecycle, last_enqueued_seq, last_acked_seq, last_pulled_seq,
-		       consecutive_failures, backoff_until, lease_owner, lease_until, last_error, updated_at
-		FROM sync_state WHERE target_key = ?`, targetKey)
-	var state SyncState
-	if err := row.Scan(&state.TargetKey, &state.Lifecycle, &state.LastEnqueuedSeq, &state.LastAckedSeq, &state.LastPulledSeq, &state.ConsecutiveFailures, &state.BackoffUntil, &state.LeaseOwner, &state.LeaseUntil, &state.LastError, &state.UpdatedAt); err != nil {
-		return nil, err
-	}
-	return &state, nil
-}
-
-func (s *Store) getSyncStateTx(tx *sql.Tx, targetKey string) (*SyncState, error) {
-	if _, err := s.execHook(tx,
-		`INSERT OR IGNORE INTO sync_state (target_key, lifecycle, updated_at) VALUES (?, ?, datetime('now'))`,
-		targetKey, SyncLifecycleIdle,
-	); err != nil {
-		return nil, err
-	}
-	row := tx.QueryRow(`
-		SELECT target_key, lifecycle, last_enqueued_seq, last_acked_seq, last_pulled_seq,
-		       consecutive_failures, backoff_until, lease_owner, lease_until, last_error, updated_at
-		FROM sync_state WHERE target_key = ?`, targetKey)
-	var state SyncState
-	if err := row.Scan(&state.TargetKey, &state.Lifecycle, &state.LastEnqueuedSeq, &state.LastAckedSeq, &state.LastPulledSeq, &state.ConsecutiveFailures, &state.BackoffUntil, &state.LeaseOwner, &state.LeaseUntil, &state.LastError, &state.UpdatedAt); err != nil {
-		return nil, err
-	}
-	return &state, nil
-}
-
-func (s *Store) backfillProjectSyncMutationsTx(tx *sql.Tx, project string) error {
-	if err := s.backfillSessionSyncMutationsTx(tx, project); err != nil {
-		return err
-	}
-	if err := s.backfillObservationSyncMutationsTx(tx, project); err != nil {
-		return err
-	}
-	return s.backfillPromptSyncMutationsTx(tx, project)
-}
-
-func (s *Store) repairEnrolledProjectSyncMutations() error {
-	return s.withTx(func(tx *sql.Tx) error {
-		rows, err := s.queryItHook(tx,
-			`SELECT project FROM sync_enrolled_projects ORDER BY project ASC`,
-		)
-		if err != nil {
-			return err
-		}
-		defer rows.Close()
-
-		var projects []string
-		for rows.Next() {
-			var project string
-			if err := rows.Scan(&project); err != nil {
-				return err
-			}
-			projects = append(projects, project)
-		}
-		if err := rows.Err(); err != nil {
-			return err
-		}
-
-		for _, project := range projects {
-			if err := s.backfillProjectSyncMutationsTx(tx, project); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-}
-
-func (s *Store) backfillSessionSyncMutationsTx(tx *sql.Tx, project string) error {
-	rows, err := s.queryItHook(tx, `
-		SELECT id, project, directory, ended_at, summary
-		FROM sessions
-		WHERE project = ?
-		  AND NOT EXISTS (
-			SELECT 1
-			FROM sync_mutations sm
-			WHERE sm.target_key = ?
-			  AND sm.entity = ?
-			  AND sm.entity_key = sessions.id
-			  AND sm.source = ?
-		  )
-		ORDER BY started_at ASC, id ASC`,
-		project, DefaultSyncTargetKey, SyncEntitySession, SyncSourceLocal,
-	)
-	if err != nil {
-		return err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var payload syncSessionPayload
-		if err := rows.Scan(&payload.ID, &payload.Project, &payload.Directory, &payload.EndedAt, &payload.Summary); err != nil {
-			return err
-		}
-		var sess Session
-		sess.ID = payload.ID
-		s.loadTagsForSession(&sess)
-		tags := sess.Tags
-		if tags == nil {
-			tags = []string{}
-		}
-		payload.Tags = &tags
-		if err := s.enqueueSyncMutationTx(tx, SyncEntitySession, payload.ID, SyncOpUpsert, payload); err != nil {
-			return err
-		}
-	}
-	return rows.Err()
-}
-
-func (s *Store) backfillObservationSyncMutationsTx(tx *sql.Tx, project string) error {
-	rows, err := s.queryItHook(tx, `
-		SELECT id, sync_id, session_id, type, title, content, tool_name, project, scope, topic_key
-		FROM observations
-		WHERE ifnull(project, '') = ?
-		  AND deleted_at IS NULL
-		  AND NOT EXISTS (
-			SELECT 1
-			FROM sync_mutations sm
-			WHERE sm.target_key = ?
-			  AND sm.entity = ?
-			  AND sm.entity_key = observations.sync_id
-			  AND sm.source = ?
-		  )
-		ORDER BY id ASC`,
-		project, DefaultSyncTargetKey, SyncEntityObservation, SyncSourceLocal,
-	)
-	if err != nil {
-		return err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var obsID int64
-		var payload syncObservationPayload
-		if err := rows.Scan(&obsID, &payload.SyncID, &payload.SessionID, &payload.Type, &payload.Title, &payload.Content, &payload.ToolName, &payload.Project, &payload.Scope, &payload.TopicKey); err != nil {
-			return err
-		}
-		var o Observation
-		o.ID = obsID
-		s.loadTagsForObservationTx(tx, &o)
-		tags := o.Tags
-		if tags == nil {
-			tags = []string{}
-		}
-		payload.Tags = &tags
-		if err := s.enqueueSyncMutationTx(tx, SyncEntityObservation, payload.SyncID, SyncOpUpsert, payload); err != nil {
-			return err
-		}
-	}
-	return rows.Err()
-}
-
-func (s *Store) backfillPromptSyncMutationsTx(tx *sql.Tx, project string) error {
-	rows, err := s.queryItHook(tx, `
-		SELECT sync_id, session_id, content, project
-		FROM user_prompts
-		WHERE ifnull(project, '') = ?
-		  AND NOT EXISTS (
-			SELECT 1
-			FROM sync_mutations sm
-			WHERE sm.target_key = ?
-			  AND sm.entity = ?
-			  AND sm.entity_key = user_prompts.sync_id
-			  AND sm.source = ?
-		  )
-		ORDER BY id ASC`,
-		project, DefaultSyncTargetKey, SyncEntityPrompt, SyncSourceLocal,
-	)
-	if err != nil {
-		return err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var payload syncPromptPayload
-		if err := rows.Scan(&payload.SyncID, &payload.SessionID, &payload.Content, &payload.Project); err != nil {
-			return err
-		}
-		if err := s.enqueueSyncMutationTx(tx, SyncEntityPrompt, payload.SyncID, SyncOpUpsert, payload); err != nil {
-			return err
-		}
-	}
-	return rows.Err()
-}
-
-func (s *Store) enqueueSyncMutationTx(tx *sql.Tx, entity, entityKey, op string, payload any) error {
-	encoded, err := json.Marshal(payload)
-	if err != nil {
-		return err
-	}
-	project := extractProjectFromPayload(payload)
-	if _, err := s.execHook(tx,
-		`INSERT OR IGNORE INTO sync_state (target_key, lifecycle, updated_at) VALUES (?, ?, datetime('now'))`,
-		DefaultSyncTargetKey, SyncLifecycleIdle,
-	); err != nil {
-		return err
-	}
-	res, err := s.execHook(tx,
-		`INSERT INTO sync_mutations (target_key, entity, entity_key, op, payload, source, project)
-		 VALUES (?, ?, ?, ?, ?, ?, ?)`,
-		DefaultSyncTargetKey, entity, entityKey, op, string(encoded), SyncSourceLocal, project,
-	)
-	if err != nil {
-		return err
-	}
-	seq, err := res.LastInsertId()
-	if err != nil {
-		return err
-	}
-	_, err = s.execHook(tx,
-		`UPDATE sync_state
-		 SET lifecycle = ?, last_enqueued_seq = ?, updated_at = datetime('now')
-		 WHERE target_key = ?`,
-		SyncLifecyclePending, seq, DefaultSyncTargetKey,
-	)
-	return err
-}
-
-// extractProjectFromPayload returns the project string from a sync payload struct.
-// It handles both string and *string Project fields across all entity payload types.
-// Returns empty string if the payload has no project or project is nil.
-func extractProjectFromPayload(payload any) string {
-	switch p := payload.(type) {
-	case syncSessionPayload:
-		return p.Project
-	case syncObservationPayload:
-		if p.Project != nil {
-			return *p.Project
-		}
-		return ""
-	case syncPromptPayload:
-		if p.Project != nil {
-			return *p.Project
-		}
-		return ""
-	default:
-		// Fallback: marshal to JSON and extract $.project via json.Unmarshal.
-		data, err := json.Marshal(payload)
-		if err != nil {
-			return ""
-		}
-		var generic struct {
-			Project *string `json:"project"`
-		}
-		if err := json.Unmarshal(data, &generic); err != nil || generic.Project == nil {
-			return ""
-		}
-		return *generic.Project
-	}
-}
-
-func decodeSyncPayload(payload []byte, dest any) error {
-	trimmed := strings.TrimSpace(string(payload))
-	if trimmed == "" {
-		return fmt.Errorf("empty payload")
-	}
-	if trimmed[0] != '"' {
-		return json.Unmarshal([]byte(trimmed), dest)
-	}
-	var encoded string
-	if err := json.Unmarshal([]byte(trimmed), &encoded); err != nil {
-		return err
-	}
-	return json.Unmarshal([]byte(encoded), dest)
-}
-
-func (s *Store) getObservationTx(tx *sql.Tx, id int64) (*Observation, error) {
-	row := tx.QueryRow(
-		`SELECT id, ifnull(sync_id, '') as sync_id, session_id, type, title, content, tool_name, project,
-		        scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		 FROM observations WHERE id = ? AND deleted_at IS NULL`, id,
-	)
-	var o Observation
-	if err := row.Scan(&o.ID, &o.SyncID, &o.SessionID, &o.Type, &o.Title, &o.Content, &o.ToolName, &o.Project, &o.Scope, &o.TopicKey, &o.RevisionCount, &o.DuplicateCount, &o.LastSeenAt, &o.CreatedAt, &o.UpdatedAt, &o.DeletedAt); err != nil {
-		return nil, err
-	}
-	s.loadTagsForObservationTx(tx, &o)
-	return &o, nil
-}
-
-func (s *Store) getObservationBySyncIDTx(tx *sql.Tx, syncID string, includeDeleted bool) (*Observation, error) {
-	query := `SELECT id, ifnull(sync_id, '') as sync_id, session_id, type, title, content, tool_name, project,
-		        scope, topic_key, revision_count, duplicate_count, last_seen_at, created_at, updated_at, deleted_at
-		 FROM observations WHERE sync_id = ?`
-	if !includeDeleted {
-		query += ` AND deleted_at IS NULL`
-	}
-	query += ` ORDER BY id DESC LIMIT 1`
-	row := tx.QueryRow(query, syncID)
-	var o Observation
-	if err := row.Scan(&o.ID, &o.SyncID, &o.SessionID, &o.Type, &o.Title, &o.Content, &o.ToolName, &o.Project, &o.Scope, &o.TopicKey, &o.RevisionCount, &o.DuplicateCount, &o.LastSeenAt, &o.CreatedAt, &o.UpdatedAt, &o.DeletedAt); err != nil {
-		return nil, err
-	}
-	return &o, nil
-}
-
-func observationPayloadFromObservation(obs *Observation) syncObservationPayload {
-	// Always include a non-nil Tags pointer so receivers can distinguish
-	// "no tags" (replace with empty) from "unknown" (nil = don't touch).
-	tags := obs.Tags
-	if tags == nil {
-		tags = []string{}
-	}
-	return syncObservationPayload{
-		SyncID:    obs.SyncID,
-		SessionID: obs.SessionID,
-		Type:      obs.Type,
-		Title:     obs.Title,
-		Content:   obs.Content,
-		ToolName:  obs.ToolName,
-		Project:   obs.Project,
-		Scope:     obs.Scope,
-		TopicKey:  obs.TopicKey,
-		Tags:      &tags,
-	}
-}
-
-func (s *Store) applySessionPayloadTx(tx *sql.Tx, payload syncSessionPayload) error {
-	if _, err := s.execHook(tx,
-		`INSERT INTO sessions (id, project, directory, ended_at, summary)
-		 VALUES (?, ?, ?, ?, ?)
-		 ON CONFLICT(id) DO UPDATE SET
-		   project = excluded.project,
-		   directory = excluded.directory,
-		   ended_at = COALESCE(excluded.ended_at, sessions.ended_at),
-		   summary = COALESCE(excluded.summary, sessions.summary)`,
-		payload.ID, payload.Project, payload.Directory, payload.EndedAt, payload.Summary,
-	); err != nil {
-		return err
-	}
-	if payload.Tags != nil {
-		if err := s.setTagsForSessionTx(tx, payload.ID, *payload.Tags); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (s *Store) applyObservationUpsertTx(tx *sql.Tx, payload syncObservationPayload) error {
-	existing, err := s.getObservationBySyncIDTx(tx, payload.SyncID, true)
-	if err == sql.ErrNoRows {
-		res, err := s.execHook(tx,
-			`INSERT INTO observations (sync_id, session_id, type, title, content, tool_name, project, scope, topic_key, normalized_hash, revision_count, duplicate_count, updated_at, deleted_at)
-			 VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 1, 1, datetime('now'), NULL)`,
-			payload.SyncID, payload.SessionID, payload.Type, payload.Title, payload.Content, payload.ToolName, payload.Project, normalizeScope(payload.Scope), payload.TopicKey, hashNormalized(payload.Content),
-		)
-		if err != nil {
-			return err
-		}
-		if payload.Tags != nil {
-			newID, err := res.LastInsertId()
-			if err != nil {
-				return err
-			}
-			if err := s.setTagsForObservationTx(tx, newID, *payload.Tags); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-	if err != nil {
-		return err
-	}
-	_, err = s.execHook(tx,
-		`UPDATE observations
-		 SET session_id = ?, type = ?, title = ?, content = ?, tool_name = ?, project = ?, scope = ?, topic_key = ?, normalized_hash = ?, revision_count = revision_count + 1, updated_at = datetime('now'), deleted_at = NULL
-		 WHERE id = ?`,
-		payload.SessionID, payload.Type, payload.Title, payload.Content, payload.ToolName, payload.Project, normalizeScope(payload.Scope), payload.TopicKey, hashNormalized(payload.Content), existing.ID,
-	)
-	if err != nil {
-		return err
-	}
-	if payload.Tags != nil {
-		if err := s.setTagsForObservationTx(tx, existing.ID, *payload.Tags); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (s *Store) applyObservationDeleteTx(tx *sql.Tx, payload syncObservationPayload) error {
-	existing, err := s.getObservationBySyncIDTx(tx, payload.SyncID, true)
-	if err == sql.ErrNoRows {
-		return nil
-	}
-	if err != nil {
-		return err
-	}
-	if payload.HardDelete {
-		_, err = s.execHook(tx, `DELETE FROM observations WHERE id = ?`, existing.ID)
-		return err
-	}
-	deletedAt := payload.DeletedAt
-	if deletedAt == nil {
-		now := Now()
-		deletedAt = &now
-	}
-	_, err = s.execHook(tx,
-		`UPDATE observations SET deleted_at = ?, updated_at = datetime('now') WHERE id = ?`,
-		deletedAt, existing.ID,
-	)
-	return err
-}
-
-func (s *Store) applyPromptUpsertTx(tx *sql.Tx, payload syncPromptPayload) error {
-	var existingID int64
-	err := tx.QueryRow(`SELECT id FROM user_prompts WHERE sync_id = ? ORDER BY id DESC LIMIT 1`, payload.SyncID).Scan(&existingID)
-	if err == sql.ErrNoRows {
-		_, err = s.execHook(tx,
-			`INSERT INTO user_prompts (sync_id, session_id, content, project) VALUES (?, ?, ?, ?)`,
-			payload.SyncID, payload.SessionID, payload.Content, payload.Project,
-		)
-		return err
-	}
-	if err != nil {
-		return err
-	}
-	_, err = s.execHook(tx,
-		`UPDATE user_prompts SET session_id = ?, content = ?, project = ? WHERE id = ?`,
-		payload.SessionID, payload.Content, payload.Project, existingID,
-	)
-	return err
-}
-
-func (s *Store) queryObservations(query string, args ...any) ([]Observation, error) {
-	rows, err := s.queryItHook(s.db, query, args...)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var results []Observation
-	for rows.Next() {
-		var o Observation
-		if err := rows.Scan(
-			&o.ID, &o.SyncID, &o.SessionID, &o.Type, &o.Title, &o.Content,
-			&o.ToolName, &o.Project, &o.Scope, &o.TopicKey, &o.RevisionCount, &o.DuplicateCount, &o.LastSeenAt,
-			&o.CreatedAt, &o.UpdatedAt, &o.DeletedAt,
-		); err != nil {
-			return nil, err
-		}
-		results = append(results, o)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	s.loadTagsForObservations(results)
-	return results, nil
-}
-
 func (s *Store) addColumnIfNotExists(tableName, columnName, definition string) error {
+	if !sqlIdentifierPattern.MatchString(tableName) || !sqlIdentifierPattern.MatchString(columnName) {
+		return fmt.Errorf("invalid migration identifier")
+	}
+	if !allowedColumnDefinitions[definition] {
+		return fmt.Errorf("unsupported migration column definition %q", definition)
+	}
 	rows, err := s.queryItHook(s.db, fmt.Sprintf("PRAGMA table_info(%s)", tableName))
 	if err != nil {
 		return err
@@ -3631,765 +653,14 @@ func (s *Store) migrateLegacyObservationsTable() error {
 	return nil
 }
 
-func nullableString(s string) *string {
-	if s == "" {
-		return nil
-	}
-	return &s
-}
-
-func truncate(s string, max int) string {
-	runes := []rune(s)
-	if len(runes) <= max {
-		return s
-	}
-	return string(runes[:max]) + "..."
-}
-
-func normalizeScope(scope string) string {
-	v := strings.TrimSpace(strings.ToLower(scope))
-	if v == "personal" {
-		return "personal"
-	}
-	return "project"
-}
-
-// SuggestTopicKey generates a stable topic key suggestion from type/title/content.
-// It infers a topic family (e.g. architecture/*, bug/*) and then appends
-// a normalized segment from title/content for stable cross-session keys.
-func SuggestTopicKey(typ, title, content string) string {
-	family := inferTopicFamily(typ, title, content)
-	cleanTitle := stripPrivateTags(title)
-	segment := normalizeTopicSegment(cleanTitle)
-
-	if segment == "" {
-		cleanContent := stripPrivateTags(content)
-		words := strings.Fields(strings.ToLower(cleanContent))
-		if len(words) > 8 {
-			words = words[:8]
-		}
-		segment = normalizeTopicSegment(strings.Join(words, " "))
-	}
-
-	if segment == "" {
-		segment = "general"
-	}
-
-	if strings.HasPrefix(segment, family+"-") {
-		segment = strings.TrimPrefix(segment, family+"-")
-	}
-	if segment == "" || segment == family {
-		segment = "general"
-	}
-
-	return family + "/" + segment
-}
-
-// SuggestTags generates tag suggestions from type/title/content.
-// It always includes the topic family as a tag, adds the normalized type when
-// it differs, and scans the text for common technical keywords.
-func SuggestTags(typ, title, content string) []string {
-	family := inferTopicFamily(typ, title, content)
-	seen := make(map[string]bool)
-	var tags []string
-
-	if family != "" && !isBlockedTag(family) {
-		seen[family] = true
-		tags = append(tags, family)
-	}
-
-	if normTyp := normalizeTag(typ); normTyp != "" && !isBlockedTag(normTyp) && !seen[normTyp] {
-		tags = append(tags, normTyp)
-		seen[normTyp] = true
-	}
-
-	text := strings.ToLower(title + " " + content)
-	tokens := make(map[string]struct{})
-	for _, tok := range strings.FieldsFunc(text, func(r rune) bool {
-		return (r < 'a' || r > 'z') && (r < '0' || r > '9')
-	}) {
-		tokens[tok] = struct{}{}
-	}
-	candidates := []string{
-		"auth", "authentication", "authorization",
-		"database", "sql", "sqlite", "postgres", "mysql",
-		"api", "rest", "http", "grpc",
-		"test", "testing",
-		"cache", "redis",
-		"config", "configuration",
-		"deploy", "deployment", "ci",
-		"security", "crypto",
-		"performance", "latency",
-		"refactor",
-		"migration",
-		"sync",
-		"session",
-		"memory",
-		"error", "panic",
-	}
-	for _, kw := range candidates {
-		// Check token presence using the raw candidate (text is lowercased, so exact match).
-		if _, ok := tokens[kw]; !ok {
-			continue
-		}
-		// Normalize and alias-resolve before appending.
-		norm := normalizeTag(kw)
-		if norm == "" || isBlockedTag(norm) || seen[norm] {
-			continue
-		}
-		tags = append(tags, norm)
-		seen[norm] = true
-		if len(tags) >= 5 {
-			break
-		}
-	}
-	return tags
-}
-
-func inferTopicFamily(typ, title, content string) string {
-	t := strings.TrimSpace(strings.ToLower(typ))
-	switch t {
-	case "architecture", "design", "adr", "refactor":
-		return "architecture"
-	case "bug", "bugfix", "fix", "incident", "hotfix":
-		return "bug"
-	case "decision":
-		return "decision"
-	case "pattern", "convention", "guideline":
-		return "pattern"
-	case "config", "setup", "infra", "infrastructure", "ci":
-		return "config"
-	case "discovery", "investigation", "root_cause", "root-cause":
-		return "discovery"
-	case "learning", "learn":
-		return "learning"
-	case "session_summary":
-		return "session"
-	}
-
-	text := strings.ToLower(title + " " + content)
-	if hasAny(text, "bug", "fix", "panic", "error", "crash", "regression", "incident", "hotfix") {
-		return "bug"
-	}
-	if hasAny(text, "architecture", "design", "adr", "boundary", "hexagonal", "refactor") {
-		return "architecture"
-	}
-	if hasAny(text, "decision", "tradeoff", "chose", "choose", "decide") {
-		return "decision"
-	}
-	if hasAny(text, "pattern", "convention", "naming", "guideline") {
-		return "pattern"
-	}
-	if hasAny(text, "config", "setup", "environment", "env", "docker", "pipeline") {
-		return "config"
-	}
-	if hasAny(text, "discovery", "investigate", "investigation", "found", "root cause") {
-		return "discovery"
-	}
-	if hasAny(text, "learned", "learning") {
-		return "learning"
-	}
-
-	if t != "" && t != "manual" {
-		return normalizeTopicSegment(t)
-	}
-
-	return "topic"
-}
-
-func hasAny(text string, words ...string) bool {
-	for _, w := range words {
-		if strings.Contains(text, w) {
-			return true
-		}
-	}
-	return false
-}
-
-func normalizeTopicSegment(s string) string {
-	v := strings.ToLower(strings.TrimSpace(s))
-	if v == "" {
-		return ""
-	}
-	v = nonAlnumRe.ReplaceAllString(v, " ")
-	v = strings.Join(strings.Fields(v), "-")
-	if len(v) > 100 {
-		v = v[:100]
-	}
-	return v
-}
-
-// tagAliases maps non-canonical tag forms to their canonical equivalents.
-// Applied after normalization, so keys must already be lowercase/hyphenated.
-var tagAliases = map[string]string{
-	"authentication": "auth",
-	"authorization":  "auth",
-	"database":       "db",
-	"databases":      "db",
-	"configuration":  "config",
-	"configurations": "config",
-	"deployment":     "deploy",
-	"deployments":    "deploy",
-	"testing":        "test",
-	"tests":          "test",
-	"errors":         "error",
-	"panics":         "panic",
-	"sessions":       "session",
-	"memories":       "memory",
-	"migrations":     "migration",
-	"caching":        "cache",
-	"caches":         "cache",
-	"performances":   "performance",
-	"securities":     "security",
-	"refactoring":    "refactor",
-	"syncing":        "sync",
-	"bugs":           "bug",
-	"fixes":          "bug",
-	"fix":            "bug",
-	"apis":           "api",
-	"configs":        "config",
-	"incidents":      "incident",
-	"patterns":       "pattern",
-	"decisions":      "decision",
-}
-
-// blockedTags contains tags that are too generic to be useful.
-// They are filtered out at storage time and never suggested.
-var blockedTags = map[string]bool{
-	"topic":  true,
-	"general": true,
-	"other":  true,
-	"misc":   true,
-	"stuff":  true,
-	"thing":  true,
-	"things": true,
-	"item":   true,
-	"items":  true,
-	"note":   true,
-	"notes":  true,
-	"update": true,
-	"change": true,
-}
-
-const (
-	// maxTagsPerObservation limits how many tags one observation can hold.
-	maxTagsPerObservation = 10
-	// maxTagsPerSession limits how many tags one session can hold.
-	maxTagsPerSession = 5
-)
-
-// isBlockedTag reports whether a (normalized) tag is too generic to store.
-func isBlockedTag(tag string) bool {
-	return blockedTags[tag]
-}
-
-// normalizeTagBase lowercases, trims, and cleans a tag string without applying
-// alias resolution. Used when the raw stored value must be matched (e.g. MergeTags).
-func normalizeTagBase(s string) string {
-	v := strings.TrimSpace(strings.ToLower(s))
-	if v == "" {
-		return ""
-	}
-	v = nonAlnumRe.ReplaceAllString(v, "-")
-	v = strings.Trim(v, "-")
-	if len(v) < 2 {
-		return ""
-	}
-	if len(v) > 50 {
-		v = v[:50]
-	}
-	return v
-}
-
-// normalizeTag normalizes a tag and resolves it to its canonical alias form.
-func normalizeTag(s string) string {
-	v := normalizeTagBase(s)
-	if v == "" {
-		return ""
-	}
-	if canonical, ok := tagAliases[v]; ok {
-		return canonical
-	}
-	return v
-}
-
-// NormalizeTag is the exported equivalent of normalizeTag, available to callers
-// outside this package (e.g. the MCP handler layer).
-func NormalizeTag(s string) string { return normalizeTag(s) }
-
-// setTagsForObservationTx replaces all tags for an observation within a
-// transaction. An empty slice removes all tags.
-func (s *Store) setTagsForObservationTx(tx *sql.Tx, obsID int64, tags []string) error {
-	if _, err := s.execHook(tx, `DELETE FROM observation_tags WHERE observation_id = ?`, obsID); err != nil {
+func (s *Store) withTx(fn func(tx *sql.Tx) error) error {
+	tx, err := s.beginTxHook()
+	if err != nil {
 		return err
 	}
-	seen := make(map[string]bool)
-	count := 0
-	for _, tag := range tags {
-		norm := normalizeTag(tag)
-		if norm == "" || isBlockedTag(norm) || seen[norm] {
-			continue
-		}
-		if count >= maxTagsPerObservation {
-			break
-		}
-		seen[norm] = true
-		count++
-		if _, err := s.execHook(tx, `INSERT OR IGNORE INTO observation_tags (observation_id, tag) VALUES (?, ?)`, obsID, norm); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// loadTagsForObservations batch-loads tags for a slice of observations and
-// assigns them in-place. Errors are silently ignored to keep reads non-fatal.
-func (s *Store) loadTagsForObservations(obs []Observation) {
-	if len(obs) == 0 {
-		return
-	}
-	placeholders := strings.Repeat("?,", len(obs))
-	placeholders = placeholders[:len(placeholders)-1]
-	args := make([]any, len(obs))
-	for i := range obs {
-		args[i] = obs[i].ID
-	}
-	rows, err := s.queryItHook(s.db,
-		fmt.Sprintf("SELECT observation_id, tag FROM observation_tags WHERE observation_id IN (%s) ORDER BY observation_id, tag", placeholders),
-		args...,
-	)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	tagMap := make(map[int64][]string)
-	for rows.Next() {
-		var obsID int64
-		var tag string
-		if rows.Scan(&obsID, &tag) == nil {
-			tagMap[obsID] = append(tagMap[obsID], tag)
-		}
-	}
-	for i := range obs {
-		if tags, ok := tagMap[obs[i].ID]; ok {
-			obs[i].Tags = tags
-		}
-	}
-}
-
-// loadTagsForObservationTx loads tags for a single observation within a transaction.
-func (s *Store) loadTagsForObservationTx(tx *sql.Tx, o *Observation) {
-	rows, err := tx.Query(
-		"SELECT tag FROM observation_tags WHERE observation_id = ? ORDER BY tag", o.ID)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var tag string
-		if rows.Scan(&tag) == nil {
-			o.Tags = append(o.Tags, tag)
-		}
-	}
-}
-
-// loadTagsForObservation loads tags for a single observation in-place.
-func (s *Store) loadTagsForObservation(o *Observation) {
-	rows, err := s.queryItHook(s.db,
-		"SELECT tag FROM observation_tags WHERE observation_id = ? ORDER BY tag", o.ID)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var tag string
-		if rows.Scan(&tag) == nil {
-			o.Tags = append(o.Tags, tag)
-		}
-	}
-}
-
-func (s *Store) setTagsForSessionTx(tx *sql.Tx, sessionID string, tags []string) error {
-	if _, err := s.execHook(tx, "DELETE FROM session_tags WHERE session_id = ?", sessionID); err != nil {
+	defer tx.Rollback()
+	if err := fn(tx); err != nil {
 		return err
 	}
-	seen := make(map[string]bool)
-	count := 0
-	for _, raw := range tags {
-		tag := normalizeTag(raw)
-		if tag == "" || isBlockedTag(tag) || seen[tag] {
-			continue
-		}
-		if count >= maxTagsPerSession {
-			break
-		}
-		seen[tag] = true
-		count++
-		if _, err := s.execHook(tx, "INSERT OR IGNORE INTO session_tags (session_id, tag) VALUES (?, ?)", sessionID, tag); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (s *Store) loadTagsForSession(sess *Session) {
-	rows, err := s.queryItHook(s.db,
-		"SELECT tag FROM session_tags WHERE session_id = ? ORDER BY tag", sess.ID)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var tag string
-		if rows.Scan(&tag) == nil {
-			sess.Tags = append(sess.Tags, tag)
-		}
-	}
-}
-
-func (s *Store) loadTagsForSessionTx(tx *sql.Tx, sess *Session) {
-	rows, err := tx.Query("SELECT tag FROM session_tags WHERE session_id = ? ORDER BY tag", sess.ID)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var tag string
-		if rows.Scan(&tag) == nil {
-			sess.Tags = append(sess.Tags, tag)
-		}
-	}
-}
-
-// SetSessionTags replaces the tag set for a session and enqueues a sync mutation.
-func (s *Store) SetSessionTags(id string, tags []string) error {
-	return s.withTx(func(tx *sql.Tx) error {
-		if err := s.setTagsForSessionTx(tx, id, tags); err != nil {
-			return err
-		}
-		var project, directory string
-		var endedAt, summary *string
-		if err := tx.QueryRow(
-			`SELECT project, directory, ended_at, summary FROM sessions WHERE id = ?`, id,
-		).Scan(&project, &directory, &endedAt, &summary); err != nil {
-			return err
-		}
-		normalizedTags := make([]string, 0, len(tags))
-		for _, raw := range tags {
-			if t := normalizeTag(raw); t != "" {
-				normalizedTags = append(normalizedTags, t)
-			}
-		}
-		return s.enqueueSyncMutationTx(tx, SyncEntitySession, id, SyncOpUpsert, syncSessionPayload{
-			ID:        id,
-			Project:   project,
-			Directory: directory,
-			EndedAt:   endedAt,
-			Summary:   summary,
-			Tags:      &normalizedTags,
-		})
-	})
-}
-
-// loadTagsForSearchResults is the SearchResult equivalent of loadTagsForObservations.
-func (s *Store) loadTagsForSearchResults(results []SearchResult) {
-	if len(results) == 0 {
-		return
-	}
-	placeholders := strings.Repeat("?,", len(results))
-	placeholders = placeholders[:len(placeholders)-1]
-	args := make([]any, len(results))
-	for i := range results {
-		args[i] = results[i].ID
-	}
-	rows, err := s.queryItHook(s.db,
-		fmt.Sprintf("SELECT observation_id, tag FROM observation_tags WHERE observation_id IN (%s) ORDER BY observation_id, tag", placeholders),
-		args...,
-	)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	tagMap := make(map[int64][]string)
-	for rows.Next() {
-		var obsID int64
-		var tag string
-		if rows.Scan(&obsID, &tag) == nil {
-			tagMap[obsID] = append(tagMap[obsID], tag)
-		}
-	}
-	for i := range results {
-		if tags, ok := tagMap[results[i].ID]; ok {
-			results[i].Tags = tags
-		}
-	}
-}
-
-func normalizeTopicKey(topic string) string {
-	v := strings.TrimSpace(strings.ToLower(topic))
-	if v == "" {
-		return ""
-	}
-	v = strings.Join(strings.Fields(v), "-")
-	if len(v) > 120 {
-		v = v[:120]
-	}
-	return v
-}
-
-func derefString(v *string) string {
-	if v == nil {
-		return ""
-	}
-	return *v
-}
-
-func hashNormalized(content string) string {
-	normalized := strings.ToLower(strings.Join(strings.Fields(content), " "))
-	h := sha256.Sum256([]byte(normalized))
-	return hex.EncodeToString(h[:])
-}
-
-func dedupeWindowExpression(window time.Duration) string {
-	if window <= 0 {
-		window = 15 * time.Minute
-	}
-	minutes := int(window.Minutes())
-	if minutes < 1 {
-		minutes = 1
-	}
-	return "-" + strconv.Itoa(minutes) + " minutes"
-}
-
-func maxInt(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
-}
-
-func normalizeSyncTargetKey(targetKey string) string {
-	if strings.TrimSpace(targetKey) == "" {
-		return DefaultSyncTargetKey
-	}
-	return strings.TrimSpace(strings.ToLower(targetKey))
-}
-
-func newSyncID(prefix string) string {
-	b := make([]byte, 8)
-	if _, err := rand.Read(b); err != nil {
-		return fmt.Sprintf("%s-%d", prefix, time.Now().UTC().UnixNano())
-	}
-	return prefix + "-" + hex.EncodeToString(b)
-}
-
-func normalizeExistingSyncID(existing, prefix string) string {
-	if strings.TrimSpace(existing) != "" {
-		return existing
-	}
-	return newSyncID(prefix)
-}
-
-// nonAlnumRe matches one or more non-alphanumeric characters. Used by
-// normalizeTagBase and normalizeTopicSegment; compiled once at package level.
-var nonAlnumRe = regexp.MustCompile(`[^a-z0-9]+`)
-
-// privateTagRegex matches <private>...</private> tags and their contents.
-// Supports multiline and nested content. Case-insensitive.
-var privateTagRegex = regexp.MustCompile(`(?is)<private>.*?</private>`)
-
-// stripPrivateTags removes all <private>...</private> content from a string.
-// This ensures sensitive information (API keys, passwords, personal data)
-// is never persisted to the memory database.
-func stripPrivateTags(s string) string {
-	result := privateTagRegex.ReplaceAllString(s, "[REDACTED]")
-	// Clean up multiple consecutive [REDACTED] and excessive whitespace
-	result = strings.TrimSpace(result)
-	return result
-}
-
-// sanitizeFTS wraps each word in quotes so FTS5 doesn't choke on special chars.
-// "fix auth bug" → `"fix" "auth" "bug"`
-func sanitizeFTS(query string) string {
-	words := strings.Fields(query)
-	for i, w := range words {
-		// Strip existing quotes to avoid double-quoting
-		w = strings.Trim(w, `"`)
-		words[i] = `"` + w + `"`
-	}
-	return strings.Join(words, " OR ")
-}
-
-// ─── Passive Capture ─────────────────────────────────────────────────────────
-
-// PassiveCaptureParams holds the input for passive memory capture.
-type PassiveCaptureParams struct {
-	SessionID string `json:"session_id"`
-	Content   string `json:"content"`
-	Project   string `json:"project,omitempty"`
-	Source    string `json:"source,omitempty"` // e.g. "subagent-stop", "session-end"
-}
-
-// PassiveCaptureResult holds the output of passive memory capture.
-type PassiveCaptureResult struct {
-	Extracted     int      `json:"extracted"`      // Total learnings found in text
-	Saved         int      `json:"saved"`          // New observations created
-	Duplicates    int      `json:"duplicates"`     // Skipped because already existed
-	SuggestedTags []string `json:"suggested_tags"` // Tag suggestions derived from the content
-}
-
-// learningHeaderPattern matches section headers for learnings in both English and Spanish.
-var learningHeaderPattern = regexp.MustCompile(
-	`(?im)^#{2,3}\s+(?:Aprendizajes(?:\s+Clave)?|Key\s+Learnings?|Learnings?):?\s*$`,
-)
-
-const (
-	minLearningLength = 20
-	minLearningWords  = 4
-)
-
-// ExtractLearnings parses structured learning items from text.
-// It looks for sections like "## Key Learnings:" or "## Aprendizajes Clave:"
-// and extracts numbered (1. text) or bullet (- text) items.
-// Returns learnings from the LAST matching section (most recent output).
-func ExtractLearnings(text string) []string {
-	matches := learningHeaderPattern.FindAllStringIndex(text, -1)
-	if len(matches) == 0 {
-		return nil
-	}
-
-	// Process sections in reverse — use first valid one (most recent)
-	for i := len(matches) - 1; i >= 0; i-- {
-		sectionStart := matches[i][1]
-		sectionText := text[sectionStart:]
-
-		// Cut off at next major section header
-		if nextHeader := regexp.MustCompile(`\n#{1,3} `).FindStringIndex(sectionText); nextHeader != nil {
-			sectionText = sectionText[:nextHeader[0]]
-		}
-
-		var learnings []string
-
-		// Try numbered items: "1. text" or "1) text"
-		numbered := regexp.MustCompile(`(?m)^\s*\d+[.)]\s+(.+)`).FindAllStringSubmatch(sectionText, -1)
-		if len(numbered) > 0 {
-			for _, m := range numbered {
-				cleaned := cleanMarkdown(m[1])
-				if len(cleaned) >= minLearningLength && len(strings.Fields(cleaned)) >= minLearningWords {
-					learnings = append(learnings, cleaned)
-				}
-			}
-		}
-
-		// Fall back to bullet items: "- text" or "* text"
-		if len(learnings) == 0 {
-			bullets := regexp.MustCompile(`(?m)^\s*[-*]\s+(.+)`).FindAllStringSubmatch(sectionText, -1)
-			for _, m := range bullets {
-				cleaned := cleanMarkdown(m[1])
-				if len(cleaned) >= minLearningLength && len(strings.Fields(cleaned)) >= minLearningWords {
-					learnings = append(learnings, cleaned)
-				}
-			}
-		}
-
-		if len(learnings) > 0 {
-			return learnings
-		}
-	}
-
-	return nil
-}
-
-// cleanMarkdown strips basic markdown formatting and collapses whitespace.
-func cleanMarkdown(text string) string {
-	text = regexp.MustCompile(`\*\*([^*]+)\*\*`).ReplaceAllString(text, "$1") // bold
-	text = regexp.MustCompile("`([^`]+)`").ReplaceAllString(text, "$1")       // inline code
-	text = regexp.MustCompile(`\*([^*]+)\*`).ReplaceAllString(text, "$1")     // italic
-	return strings.TrimSpace(strings.Join(strings.Fields(text), " "))
-}
-
-// PassiveCapture extracts learnings from text and saves them as observations.
-// It deduplicates against existing observations using content hash matching.
-func (s *Store) PassiveCapture(p PassiveCaptureParams) (*PassiveCaptureResult, error) {
-	result := &PassiveCaptureResult{}
-
-	learnings := ExtractLearnings(p.Content)
-	result.Extracted = len(learnings)
-
-	if len(learnings) == 0 {
-		return result, nil
-	}
-
-	seenTags := make(map[string]bool)
-	for _, learning := range learnings {
-		for _, tag := range SuggestTags("passive", learning, learning) {
-			if !seenTags[tag] {
-				result.SuggestedTags = append(result.SuggestedTags, tag)
-				seenTags[tag] = true
-			}
-		}
-	}
-
-	for _, learning := range learnings {
-		// Check if this learning already exists (by content hash) within this project
-		normHash := hashNormalized(learning)
-		var existingID int64
-		err := s.db.QueryRow(
-			`SELECT id FROM observations
-			 WHERE normalized_hash = ?
-			   AND ifnull(project, '') = ifnull(?, '')
-			   AND deleted_at IS NULL
-			 LIMIT 1`,
-			normHash, nullableString(p.Project),
-		).Scan(&existingID)
-
-		if err == nil {
-			// Already exists — skip
-			result.Duplicates++
-			continue
-		}
-
-		// Truncate for title: first 60 chars
-		title := learning
-		if len(title) > 60 {
-			title = title[:60] + "..."
-		}
-
-		_, err = s.AddObservation(AddObservationParams{
-			SessionID: p.SessionID,
-			Type:      "passive",
-			Title:     title,
-			Content:   learning,
-			Project:   p.Project,
-			Scope:     "project",
-			ToolName:  p.Source,
-		})
-		if err != nil {
-			return result, fmt.Errorf("passive capture save: %w", err)
-		}
-		result.Saved++
-	}
-
-	return result, nil
-}
-
-// ClassifyTool returns the observation type for a given tool name.
-func ClassifyTool(toolName string) string {
-	switch toolName {
-	case "write", "edit", "patch":
-		return "file_change"
-	case "bash":
-		return "command"
-	case "read", "view":
-		return "file_read"
-	case "grep", "glob", "ls":
-		return "search"
-	default:
-		return "tool_use"
-	}
-}
-
-// Now returns the current time formatted for SQLite.
-func Now() string {
-	return time.Now().UTC().Format("2006-01-02 15:04:05")
+	return s.commitHook(tx)
 }
