+++ sqlc.yaml
@@ -1,0 +1,13 @@
+version: "2"
+sql:
+  - engine: "sqlite"
+    schema: "internal/db/schema.sql"
+    queries: "internal/db/queries"
+    gen:
+      go:
+        package: "db"
+        out: "internal/db/generated"
+        emit_interface: true
+        emit_json_tags: true
+        emit_empty_slices: true
+
